Module vmm::arch

Implements platform specific functionality. Supported platforms: x86_64 and aarch64.

Re-exports

pub use aarch64::arch_memory_regions;

pub use aarch64::configure_system;

pub use aarch64::get_kernel_start;

pub use aarch64::initrd_load_addr;

pub use aarch64::layout::CMDLINE_MAX_SIZE;

pub use aarch64::layout::IRQ_BASE;

pub use aarch64::layout::IRQ_MAX;

pub use aarch64::ConfigurationError;

pub use aarch64::MMIO_MEM_SIZE;

pub use aarch64::MMIO_MEM_START;

Modules

aarch64

Module for aarch64 related functionality.

Structs

InitrdConfig

Type for passing information about the initrd in the guest memory.

Enums

DeviceType

Types of devices that can get attached to this platform.

Constants

PAGE_SIZE

Default (smallest) memory page size for the supported architectures.