Module vmm::arch::x86_64

Module for x86_64 related functionality.

Modules

cpu_model

Logic for handling x86_64 CPU models.

gdt 🔒

interrupts

Contains logic for setting up Advanced Programmable Interrupt Controller (local version).

layout

Layout for the x86_64 system.

mptable 🔒

msr

Logic for configuring x86_64 model specific registers (MSRs).

regs

Logic for configuring x86_64 registers.

Enums

ConfigurationError

Errors thrown while configuring x86_64 system.

Constants

E820_RAM 🔒

E820_RESERVED 🔒

EBDA_SIZE 🔒

EBDA_START 🔒

FIRST_ADDR_PAST_32BITS 🔒

MEM_32BIT_GAP_SIZE

Size of MMIO gap at top of 32-bit address space.

MMIO_MEM_SIZE

The size of the memory area reserved for MMIO devices.

MMIO_MEM_START

The start of the memory area reserved for MMIO devices.

_DERIVE_Display_FOR_ConfigurationError 🔒

Functions

add_e820_entry 🔒

Add an e820 region to the e820 map. Returns Ok(()) if successful, or an error if there is no space left in the map.

arch_memory_regions

Returns a Vec of the valid memory addresses. These should be used to configure the GuestMemoryMmap structure for the platform. For x86_64 all addresses are valid from the start of the kernel except a carve out at the end of 32bit address space.

configure_system

Configures the system and should be called once per vm before starting vcpu threads.

get_kernel_start

Returns the memory address where the kernel could be loaded.

initrd_load_addr

Returns the memory address where the initrd could be loaded.