Module vmm::builder

Handles setup and initialization a Vmm object. Enables pre-boot setup, instantiation and booting of a Firecracker VMM.

Enums

BuildMicrovmFromSnapshotError

Error type for build_microvm_from_snapshot.

StartMicrovmError

Errors associated with starting the instance.

Constants

_DERIVE_Display_FOR_BuildMicrovmFromSnapshotError 🔒

_DERIVE_Display_FOR_StartMicrovmError 🔒

Functions

attach_balloon_device 🔒

attach_block_devices 🔒

attach_boot_timer_device 🔒

attach_entropy_device 🔒

attach_net_devices 🔒

attach_unixsock_vsock_device 🔒

attach_virtio_device 🔒

Attaches a VirtioDevice device to the device manager and event manager.

build_and_boot_microvm

Builds and boots a microVM based on the current Firecracker VmResources configuration.

build_microvm_for_boot

Builds and starts a microVM based on the current Firecracker VmResources configuration.

build_microvm_from_snapshot

Builds and starts a microVM based on the provided MicrovmState.

configure_system_for_boot

Configures the system for booting Linux.

create_vcpus 🔒

create_vmm_and_vcpus 🔒

load_initrd 🔒

Loads the initrd from a file into the given memory slice.

load_initrd_from_config 🔒

load_kernel 🔒

set_stdout_nonblocking 🔒

setup_interrupt_controller

Sets up the irqchip for a x86_64 microVM.

setup_serial_device

Sets up the serial device.