Module vmm::cpu_config::x86_64::cpuid

Module for CPUID instruction related content Utility for configuring the CPUID (CPU identification) for the guest microVM.

Re-exports

pub use amd::AmdCpuid;

pub use intel::IntelCpuid;

Modules

amd

AMD CPUID specification handling.

common

cpuid utility functions.

intel

Intel CPUID specification handling.

normalize 🔒

CPUID normalize implementation.

Structs

CpuidEntry

CPUID entry information stored for each leaf of IntelCpuid.

CpuidKey

CPUID index values leaf and subleaf.

CpuidRegisters

To transmute this into leaves such that we can return mutable reference to it with leaf specific accessors, requires this to have a consistent member ordering. core::arch::x86_64::CpuidResult is not repr(C).

KvmCpuidFlags

Definitions from `kvm/arch/x86/include/uapi/asm/kvm.h

MissingBrandStringLeaves

Error type for CpuidTrait::apply_brand_string.

Enums

Cpuid

CPUID information

CpuidTryFromKvmCpuid

Error type for conversion from kvm_bindings::CpuId to Cpuid.

FeatureInformationError

Error type for setting leaf 1 section of IntelCpuid::normalize.

GetMaxCpusPerPackageError

Error type for get_max_cpus_per_package.

NormalizeCpuidError

Error type for super::Cpuid::normalize.

Constants

BRAND_STRING_LENGTH

To store the brand string we have 3 leaves, each with 4 registers, each with 4 bytes.

VENDOR_ID_AMD

AMD brand string.

VENDOR_ID_AMD_STR

AMD brand string.

VENDOR_ID_INTEL

Intel brand string.

VENDOR_ID_INTEL_STR

Intel brand string.

_DERIVE_Display_FOR_CpuidTryFromKvmCpuid 🔒

Traits

CpuidTrait

Trait defining shared behaviour between CPUID structures.

Functions

cpuid 🔒

Mimic of std::arch::x86_64::__cpuid that wraps cpuid_count.

cpuid_count 🔒

Safe wrapper around std::arch::x86_64::__cpuid_count.

host_brand_string

Gets the Intel default brand. Gets host brand string.