Module vmm::cpu_config::x86_64::custom_cpu_template

Module for custom CPU templates

Structs

CpuidLeafModifier

Composite type that holistically provides the location of a specific register being used in the context of a CPUID tree.

CpuidRegisterModifier

Target register to be modified by a bitmap.

CustomCpuTemplate

Wrapper type to containing x86_64 CPU config modifiers.

RegisterModifier

Wrapper of a mask defined as a bitmap to apply changes to a given register’s value.

Enums

CpuidRegister

CPUID register enumeration

Functions

deserialize_cpuid_register 🔒

deserialize_kvm_cpuid_flags 🔒

serialize_cpuid_register 🔒