Struct vmm::device_manager::mmio::MMIODeviceManager

pub struct MMIODeviceManager {
    pub(crate) bus: Bus,
    pub(crate) irq_allocator: IdAllocator,
    pub(crate) address_allocator: AddressAllocator,
    pub(crate) id_to_dev_info: HashMap<(DeviceType, String), MMIODeviceInfo>,
}

Manages the complexities of registering a MMIO device.

Fields
bus: Busirq_allocator: IdAllocatoraddress_allocator: AddressAllocatorid_to_dev_info: HashMap<(DeviceType, String), MMIODeviceInfo>
Implementations

impl MMIODeviceManager

pub fn new( mmio_base: u64, mmio_size: u64, (irq_start, irq_end): (u32, u32) ) -> Result<MMIODeviceManager, MmioError>

Create a new DeviceManager handling mmio devices (virtio net, block).

fn allocate_mmio_resources( &mut self, irq_count: u32 ) -> Result<MMIODeviceInfo, MmioError>

Allocates resources for a new device to be added.

fn register_mmio_device( &mut self, identifier: (DeviceType, String), device_info: MMIODeviceInfo, device: Arc<Mutex<BusDevice>> ) -> Result<(), MmioError>

Register a device at some MMIO address.

pub fn register_mmio_virtio( &mut self, vm: &VmFd, device_id: String, mmio_device: MmioTransport, device_info: &MMIODeviceInfo ) -> Result<(), MmioError>

Register a virtio-over-MMIO device to be used via MMIO transport at a specific slot.

pub fn add_virtio_device_to_cmdline( cmdline: &mut Cmdline, device_info: &MMIODeviceInfo ) -> Result<(), MmioError>

Append a registered virtio-over-MMIO device to the kernel cmdline.

pub fn register_mmio_virtio_for_boot( &mut self, vm: &VmFd, device_id: String, mmio_device: MmioTransport, _cmdline: &mut Cmdline ) -> Result<MMIODeviceInfo, MmioError>

Allocate slot and register an already created virtio-over-MMIO device. Also Adds the device to the boot cmdline.

pub fn register_mmio_boot_timer( &mut self, device: BootTimer ) -> Result<(), MmioError>

Register a boot timer device.

pub fn get_device_info(&self) -> &HashMap<(DeviceType, String), MMIODeviceInfo>

Gets the information of the devices registered up to some point in time.

pub fn get_device( &self, device_type: DeviceType, device_id: &str ) -> Option<&Mutex<BusDevice>>

Gets the specified device.

pub fn for_each_device<F, E: Debug>(&self, f: F) -> Result<(), E>
where F: FnMut(&DeviceType, &String, &MMIODeviceInfo, &Mutex<BusDevice>) -> Result<(), E>,

Run fn for each registered device.

pub fn for_each_virtio_device<F, E: Debug>(&self, f: F) -> Result<(), E>
where F: FnMut(u32, &String, &MMIODeviceInfo, Arc<Mutex<dyn VirtioDevice>>) -> Result<(), E>,

Run fn for each registered virtio device.

pub fn with_virtio_device_with_id<T, F>( &self, virtio_type: u32, id: &str, f: F ) -> Result<(), MmioError>
where T: VirtioDevice + 'static + Debug, F: FnOnce(&mut T) -> Result<(), String>,

Run fn f() for the virtio device matching virtio_type and id.

pub fn kick_devices(&self)

Artificially kick devices as if they had external events.