Module vmm::device_manager::persist

Device managers (de)serialization support. Provides functionality for saving/restoring the MMIO device manager and its devices.

Structs

ConnectedBalloonState

Holds the state of a balloon device connected to the MMIO space.

ConnectedEntropyState

Holds the state of an entropy device connected to the MMIO space.

ConnectedNetState

Holds the state of a net device connected to the MMIO space.

ConnectedVhostUserBlockState

Holds the state of a vhost-user block device connected to the MMIO space.

ConnectedVirtioBlockState

Holds the state of a virtio block device connected to the MMIO space.

ConnectedVsockState

Holds the state of a vsock device connected to the MMIO space.

DeviceStates

Holds the device states.

MMIODevManagerConstructorArgs

Enums

DevicePersistError

Errors for (de)serialization of the MMIO device manager.

MmdsVersionState

Holds the MMDS data store version.

SharedDeviceType

A type used to extract the concrete Arc<Mutex<T>> for each of the device types when restoring from a snapshot.

Constants

_DERIVE_Display_FOR_DevicePersistError 🔒