Module vmm::devices::legacy::i8042

Structs

I8042Device

A i8042 PS/2 controller that emulates just enough to shutdown the machine.

I8042DeviceMetrics 🔒

Metrics specific to the i8042 device.

Enums

I8042Error

Errors thrown by the i8042 device.

Constants

BUF_SIZE 🔒

Internal i8042 buffer size, in bytes

CB_KBD_INT 🔒

i8042 control register bits

CB_POST_OK 🔒

CMD_READ_CTR 🔒

i8042 commands These values are written by the guest driver to port 0x64.

CMD_READ_OUTP 🔒

CMD_RESET_CPU 🔒

CMD_WRITE_CTR 🔒

CMD_WRITE_OUTP 🔒

KEY_ALT 🔒

KEY_CTRL 🔒

Key scan codes

KEY_DEL 🔒

OFS_DATA 🔒

Offset of the data port (port 0x60)

OFS_STATUS 🔒

Offset of the status port (port 0x64)

SB_I8042_CMD_DATA 🔒

SB_KBD_ENABLED 🔒

SB_OUT_DATA_AVAIL 🔒

i8042 status register bits

_DERIVE_Display_FOR_I8042Error 🔒

Statics

METRICS 🔒

Stores aggregated metrics