Module vmm::devices::legacy::serial

Implements a wrapper over an UART serial device.

Structs

SerialDeviceMetrics

Metrics specific to the UART device.

SerialEventsWrapper

Wrapper over available events (i.e metrics, buffer ready etc).

SerialWrapper

Wrapper over the imported serial device.

Enums

RawIOError

SerialOut

Constants

IER_RDA_BIT

Received Data Available interrupt - for letting the driver know that there is some pending data to be processed.

IER_RDA_OFFSET

Received Data Available interrupt offset

_DERIVE_Display_FOR_RawIOError 🔒

Statics

METRICS 🔒

Stores aggregated metrics

Traits

RawIOHandler

Functions

is_fifo 🔒

Checks whether the given file descriptor is a FIFO pipe.

Type Aliases

SerialDevice

Type for representing a serial device.