Struct vmm::devices::legacy::I8042Device

pub struct I8042Device {
    reset_evt: EventFd,
    kbd_interrupt_evt: EventFd,
    status: u8,
    control: u8,
    outp: u8,
    cmd: u8,
    buf: [u8; 16],
    bhead: Wrapping<usize>,
    btail: Wrapping<usize>,
}

A i8042 PS/2 controller that emulates just enough to shutdown the machine.

Fields
reset_evt: EventFd

CPU reset eventfd. We will set this event when the guest issues CMD_RESET_CPU.

kbd_interrupt_evt: EventFd

Keyboard interrupt event (IRQ 1).

status: u8

The i8042 status register.

control: u8

The i8042 control register.

outp: u8

The i8042 output port.

cmd: u8

The last command sent to port 0x64.

buf: [u8; 16]

The internal i8042 data buffer.

bhead: Wrapping<usize>btail: Wrapping<usize>
Implementations

impl I8042Device

pub fn new(reset_evt: EventFd, kbd_interrupt_evt: EventFd) -> I8042Device

Constructs an i8042 device that will signal the given event when the guest requests it.

pub fn trigger_ctrl_alt_del(&mut self) -> Result<(), I8042Error>

Signal a ctrl-alt-del (reset) event.

fn trigger_kbd_interrupt(&self) -> Result<(), I8042Error>

fn trigger_key(&mut self, key: u16) -> Result<(), I8042Error>

fn push_byte(&mut self, byte: u8) -> Result<(), I8042Error>

fn pop_byte(&mut self) -> Option<u8>

fn flush_buf(&mut self)

fn buf_len(&self) -> usize

impl I8042Device

pub fn bus_read(&mut self, offset: u64, data: &mut [u8])

pub fn bus_write(&mut self, offset: u64, data: &[u8])