Module vmm::devices::virtio

Implements virtio devices, queues, and transport mechanisms.

Modules

balloon

Implements a virtio balloon device.

block_common

device

device_status 🔒

When the driver initializes the device, it lets the device know about the completed stages using the Device Status Field.

gen

iovec

mmio

net

Implements a virtio network device.

persist

Defines the structures needed for saving/restoring Virtio primitives.

queue

rng

vhost_user

vhost_user_block

vhost_user_metrics

Defines the metrics system for vhost-user devices.

virtio_block

Implements a virtio block device.

vsock

The Firecracker vsock device aims to provide full virtio-vsock support to software running inside the guest VM, while bypassing vhost kernel code on the host. To that end, Firecracker implements the virtio-vsock device model, and mediates communication between AF_UNIX sockets (on the host end) and AF_VSOCK sockets (on the guest end).

Enums

ActivateError

Errors triggered when activating a VirtioDevice.

Constants

NOTIFY_REG_OFFSET

Offset from the base MMIO address of a virtio device used by the guest to notify the device of queue events.

TYPE_BALLOON

Virtio balloon device ID.

TYPE_BLOCK

Virtio block device ID.

TYPE_NET

Types taken from linux/virtio_ids.h. Type 0 is not used by virtio. Use it as wildcard for non-virtio devices Virtio net device ID.

TYPE_RNG

Virtio rng device ID.

_DERIVE_Display_FOR_ActivateError 🔒

Traits

AsAny

Trait that helps in upcasting an object to Any