Trait vmm::devices::virtio::device::VirtioDevice

pub trait VirtioDevice: AsAny + Send {

    // Required methods
    fn avail_features(&self) -> u64;
    fn acked_features(&self) -> u64;
    fn set_acked_features(&mut self, acked_features: u64);
    fn device_type(&self) -> u32;
    fn queues(&self) -> &[Queue];
    fn queues_mut(&mut self) -> &mut [Queue];
    fn queue_events(&self) -> &[EventFd];
    fn interrupt_evt(&self) -> &EventFd;
    fn interrupt_status(&self) -> Arc<AtomicU32>;
    fn read_config(&self, offset: u64, data: &mut [u8]);
    fn write_config(&mut self, offset: u64, data: &[u8]);
    fn activate(&mut self, mem: GuestMemoryMmap) -> Result<(), ActivateError>;
    fn is_activated(&self) -> bool;

    // Provided methods
    fn has_feature(&self, feature: u64) -> bool { ... }
    fn avail_features_by_page(&self, page: u32) -> u32 { ... }
    fn ack_features_by_page(&mut self, page: u32, value: u32) { ... }
    fn reset(&mut self) -> Option<(EventFd, Vec<EventFd>)> { ... }

}

Trait for virtio devices to be driven by a virtio transport.

The lifecycle of a virtio device is to be moved to a virtio transport, which will then query the device. The virtio devices needs to create queues, events and event fds for interrupts and expose them to the transport via get_queues/get_queue_events/get_interrupt/get_interrupt_status fns.

Required Methods

fn avail_features(&self) -> u64

Get the available features offered by device.

fn acked_features(&self) -> u64

Get acknowledged features of the driver.

fn set_acked_features(&mut self, acked_features: u64)

Set acknowledged features of the driver. This function must maintain the following invariant:

self.avail_features() & self.acked_features() = self.get_acked_features()

fn device_type(&self) -> u32

The virtio device type.

fn queues(&self) -> &[Queue]

Returns the device queues.

fn queues_mut(&mut self) -> &mut [Queue]

Returns a mutable reference to the device queues.

fn queue_events(&self) -> &[EventFd]

Returns the device queues event fds.

fn interrupt_evt(&self) -> &EventFd

Returns the device interrupt eventfd.

fn interrupt_status(&self) -> Arc<AtomicU32>

Returns the current device interrupt status.

fn read_config(&self, offset: u64, data: &mut [u8])

Reads this device configuration space at offset.

fn write_config(&mut self, offset: u64, data: &[u8])

Writes to this device configuration space at offset.

fn activate(&mut self, mem: GuestMemoryMmap) -> Result<(), ActivateError>

Performs the formal activation for a device, which can be verified also with is_activated.

fn is_activated(&self) -> bool

Checks if the resources of this device are activated.

Provided Methods

fn has_feature(&self, feature: u64) -> bool

Check if virtio device has negotiated given feature.

fn avail_features_by_page(&self, page: u32) -> u32

The set of feature bits shifted by page * 32.

fn ack_features_by_page(&mut self, page: u32, value: u32)

Acknowledges that this set of features should be enabled.

fn reset(&mut self) -> Option<(EventFd, Vec<EventFd>)>

Optionally deactivates this device and returns ownership of the guest memory map, interrupt event, and queue events.