Struct vmm::devices::virtio::mmio::MmioTransport

pub struct MmioTransport {
    device: Arc<Mutex<dyn VirtioDevice>>,
    pub(crate) features_select: u32,
    pub(crate) acked_features_select: u32,
    pub(crate) queue_select: u32,
    pub(crate) device_status: u32,
    pub(crate) config_generation: u32,
    mem: GuestMemoryMmap,
    pub(crate) interrupt_status: Arc<AtomicU32>,
    pub is_vhost_user: bool,
}

Implements the MMIO transport for virtio devices.

This requires 3 points of installation to work with a VM:

Mmio reads and writes must be sent to this device at what is referred to here as MMIO base.

Mmio::queue_evts must be installed at virtio::NOTIFY_REG_OFFSET offset from the MMIO base. Each event in the array must be signaled if the index is written at that offset.

Mmio::interrupt_evt must signal an interrupt that the guest driver is listening to when it is written to.

Typically one page (4096 bytes) of MMIO address space is sufficient to handle this transport and inner virtio device.

Fields
device: Arc<Mutex<dyn VirtioDevice>>features_select: u32acked_features_select: u32queue_select: u32device_status: u32config_generation: u32mem: GuestMemoryMmapinterrupt_status: Arc<AtomicU32>is_vhost_user: bool
Implementations

impl MmioTransport

pub fn new( mem: GuestMemoryMmap, device: Arc<Mutex<dyn VirtioDevice>>, is_vhost_user: bool ) -> MmioTransport

Constructs a new MMIO transport for the given virtio device.

pub fn locked_device(&self) -> MutexGuard<'_, dyn VirtioDevice + 'static>

Gets the encapsulated locked VirtioDevice.

pub fn device(&self) -> Arc<Mutex<dyn VirtioDevice>>

Gets the encapsulated VirtioDevice.

fn check_device_status(&self, set: u32, clr: u32) -> bool

fn are_queues_valid(&self) -> bool

fn with_queue<U, F>(&self, d: U, f: F) -> U
where F: FnOnce(&Queue) -> U, U: Debug,

fn with_queue_mut<F: FnOnce(&mut Queue)>(&mut self, f: F) -> bool

fn update_queue_field<F: FnOnce(&mut Queue)>(&mut self, f: F)

fn reset(&mut self)

fn set_device_status(&mut self, status: u32)

Update device status according to the state machine defined by VirtIO Spec 1.0. Please refer to VirtIO Spec 1.0, section 2.1.1 and 3.1.1.

The driver MUST update device status, setting bits to indicate the completed steps of the driver initialization sequence specified in 3.1. The driver MUST NOT clear a device status bit. If the driver sets the FAILED bit, the driver MUST later reset the device before attempting to re-initialize.

impl MmioTransport

pub fn bus_read(&mut self, offset: u64, data: &mut [u8])

pub fn bus_write(&mut self, offset: u64, data: &[u8])