Module vmm::devices::virtio::net

Implements a virtio network device.

Re-exports

pub use self::device::Net;

Modules

device

event_handler 🔒

gen 🔒

metrics

Defines the metrics system for Network devices.

persist

Defines the structures needed for saving/restoring net devices.

tap 🔒

Structs

Tap

Handle for a network tap interface.

Enums

NetError

Errors the network device can trigger.

NetQueue

Enum representing the Net device queue types

TapError

List of errors the tap implementation can throw.

Constants

MAX_BUFFER_SIZE

Maximum size of the frame buffers handled by this device.

NET_NUM_QUEUES

The number of queues of the network device.

NET_QUEUE_SIZES

RX_INDEX

The index of the rx queue from Net device queues/queues_evts vector.

TX_INDEX

The index of the tx queue from Net device queues/queues_evts vector.

_DERIVE_Display_FOR_NetError 🔒