Struct vmm::devices::virtio::net::device::Net

pub struct Net {

    pub(crate) id: String,
    pub tap: Tap,
    pub(crate) avail_features: u64,
    pub(crate) acked_features: u64,
    pub(crate) queues: Vec<Queue>,
    pub(crate) queue_evts: Vec<EventFd>,
    pub(crate) rx_rate_limiter: RateLimiter,
    pub(crate) tx_rate_limiter: RateLimiter,
    pub(crate) rx_deferred_frame: bool,
    rx_bytes_read: usize,
    rx_frame_buf: [u8; 65562],
    tx_frame_headers: [u8; 54],
    pub(crate) irq_trigger: IrqTrigger,
    pub(crate) config_space: ConfigSpace,
    pub(crate) guest_mac: Option<MacAddr>,
    pub(crate) device_state: DeviceState,
    pub(crate) activate_evt: EventFd,
    pub mmds_ns: Option<MmdsNetworkStack>,
    pub(crate) metrics: Arc<NetDeviceMetrics>,

}

VirtIO network device.

It emulates a network device able to exchange L2 frames between the guest and a host-side tap device.

Fields
id: Stringtap: Tap

The backend for this device: a tap.

avail_features: u64acked_features: u64queues: Vec<Queue>queue_evts: Vec<EventFd>rx_rate_limiter: RateLimitertx_rate_limiter: RateLimiterrx_deferred_frame: boolrx_bytes_read: usizerx_frame_buf: [u8; 65562]tx_frame_headers: [u8; 54]irq_trigger: IrqTriggerconfig_space: ConfigSpaceguest_mac: Option<MacAddr>device_state: DeviceStateactivate_evt: EventFdmmds_ns: Option<MmdsNetworkStack>

The MMDS stack corresponding to this interface. Only if MMDS transport has been associated with it.

metrics: Arc<NetDeviceMetrics>
Implementations

impl Net

pub fn new_with_tap( id: String, tap: Tap, guest_mac: Option<MacAddr>, rx_rate_limiter: RateLimiter, tx_rate_limiter: RateLimiter ) -> Result<Self, NetError>

Create a new virtio network device with the given TAP interface.

pub fn new( id: String, tap_if_name: &str, guest_mac: Option<MacAddr>, rx_rate_limiter: RateLimiter, tx_rate_limiter: RateLimiter ) -> Result<Self, NetError>

Create a new virtio network device given the interface name.

pub fn id(&self) -> &String

Provides the ID of this net device.

pub fn guest_mac(&self) -> Option<&MacAddr>

Provides the MAC of this net device.

pub fn iface_name(&self) -> String

Provides the host IFACE name of this net device.

pub fn mmds_ns(&self) -> Option<&MmdsNetworkStack>

Provides the MmdsNetworkStack of this net device.

pub fn configure_mmds_network_stack( &mut self, ipv4_addr: Ipv4Addr, mmds: Arc<Mutex<Mmds>> )

Configures the MmdsNetworkStack to allow device to forward MMDS requests. If the device already supports MMDS, updates the IPv4 address.

pub fn disable_mmds_network_stack(&mut self)

Disables the MmdsNetworkStack to prevent device to forward MMDS requests.

pub fn rx_rate_limiter(&self) -> &RateLimiter

Provides a reference to the configured RX rate limiter.

pub fn tx_rate_limiter(&self) -> &RateLimiter

Provides a reference to the configured TX rate limiter.

fn signal_used_queue(&mut self, queue_type: NetQueue) -> Result<(), DeviceError>

fn rate_limiter_consume_op(rate_limiter: &mut RateLimiter, size: u64) -> bool

fn rate_limiter_replenish_op(rate_limiter: &mut RateLimiter, size: u64)

fn rate_limited_rx_single_frame(&mut self) -> bool

fn write_to_descriptor_chain( mem: &GuestMemoryMmap, data: &[u8], head: DescriptorChain<'_>, net_metrics: &NetDeviceMetrics ) -> Result<(), FrontendError>

Write a slice in a descriptor chain

Errors

Returns an error if the descriptor chain is too short or an inappropriate (read only) descriptor is found in the chain

fn do_write_frame_to_guest(&mut self) -> Result<(), FrontendError>

fn write_frame_to_guest(&mut self) -> bool

fn write_to_mmds_or_tap( mmds_ns: Option<&mut MmdsNetworkStack>, rate_limiter: &mut RateLimiter, headers: &mut [u8], frame_iovec: &IoVecBuffer, tap: &mut Tap, guest_mac: Option<MacAddr>, net_metrics: &NetDeviceMetrics ) -> Result<bool, NetError>

fn read_from_mmds_or_tap(&mut self) -> Result<usize, NetError>

fn process_rx(&mut self) -> Result<(), DeviceError>

fn handle_deferred_frame(&mut self) -> Result<(), DeviceError>

fn resume_rx(&mut self) -> Result<(), DeviceError>

fn process_tx(&mut self) -> Result<(), DeviceError>

pub fn patch_rate_limiters( &mut self, rx_bytes: BucketUpdate, rx_ops: BucketUpdate, tx_bytes: BucketUpdate, tx_ops: BucketUpdate )

Updates the parameters for the rate limiters

fn read_tap(&mut self) -> Result<usize>

fn write_tap(tap: &mut Tap, buf: &IoVecBuffer) -> Result<usize>

pub fn process_rx_queue_event(&mut self)

Process a single RX queue event.

This is called by the event manager responding to the guest adding a new buffer in the RX queue.

pub fn process_tap_rx_event(&mut self)

pub fn process_tx_queue_event(&mut self)

Process a single TX queue event.

This is called by the event manager responding to the guest adding a new buffer in the TX queue.

pub fn process_rx_rate_limiter_event(&mut self)

pub fn process_tx_rate_limiter_event(&mut self)

pub fn process_virtio_queues(&mut self)

Process device virtio queue(s).

impl Net

fn register_runtime_events(&self, ops: &mut EventOps<'_>)

fn register_activate_event(&self, ops: &mut EventOps<'_>)

fn process_activate_event(&self, ops: &mut EventOps<'_>)