Struct vmm::devices::virtio::queue::Queue

pub struct Queue {
    pub(crate) max_size: u16,
    pub size: u16,
    pub ready: bool,
    pub desc_table: GuestAddress,
    pub avail_ring: GuestAddress,
    pub used_ring: GuestAddress,
    pub(crate) next_avail: Wrapping<u16>,
    pub(crate) next_used: Wrapping<u16>,
    pub(crate) uses_notif_suppression: bool,
    pub(crate) num_added: Wrapping<u16>,
}

A virtio queue’s parameters.

Fields
max_size: u16

The maximal size in elements offered by the device

size: u16

The queue size in elements the driver selected

ready: bool

Indicates if the queue is finished with configuration

desc_table: GuestAddress

Guest physical address of the descriptor table

avail_ring: GuestAddress

Guest physical address of the available ring

used_ring: GuestAddress

Guest physical address of the used ring

next_avail: Wrapping<u16>next_used: Wrapping<u16>uses_notif_suppression: bool

VIRTIO_F_RING_EVENT_IDX negotiated (notification suppression enabled)

num_added: Wrapping<u16>

The number of added used buffers since last guest kick

Implementations

impl Queue

pub fn new(max_size: u16) -> Queue

Constructs an empty virtio queue with the given max_size.

pub fn get_max_size(&self) -> u16

Maximum size of the queue.

pub fn actual_size(&self) -> u16

Return the actual size of the queue, as the driver may not set up a queue as big as the device allows.

pub fn is_layout_valid<M: GuestMemory>(&self, mem: &M) -> bool

Validates the queue’s in-memory layout is correct.

pub fn is_valid<M: GuestMemory>(&self, mem: &M) -> bool

Validates that the queue’s representation is correct.

fn len<M: GuestMemory>(&self, mem: &M) -> u16

Returns the number of yet-to-be-popped descriptor chains in the avail ring.

pub fn is_empty<M: GuestMemory>(&self, mem: &M) -> bool

Checks if the driver has made any descriptor chains available in the avail ring.

pub fn pop<'b, M: GuestMemory>( &mut self, mem: &'b M ) -> Option<DescriptorChain<'b, M>>

Pop the first available descriptor chain from the avail ring.

pub fn pop_or_enable_notification<'b, M: GuestMemory>( &mut self, mem: &'b M ) -> Option<DescriptorChain<'b, M>>

Try to pop the first available descriptor chain from the avail ring. If no descriptor is available, enable notifications.

fn do_pop_unchecked<'b, M: GuestMemory>( &mut self, mem: &'b M ) -> Option<DescriptorChain<'b, M>>

Pop the first available descriptor chain from the avail ring.

Important

This is an internal method that ASSUMES THAT THERE ARE AVAILABLE DESCRIPTORS. Otherwise it will retrieve a descriptor that contains garbage data (obsolete/empty).

pub fn undo_pop(&mut self)

Undo the effects of the last self.pop() call. The caller can use this, if it was unable to consume the last popped descriptor chain.

pub fn add_used<M: GuestMemory>( &mut self, mem: &M, desc_index: u16, len: u32 ) -> Result<(), QueueError>

Puts an available descriptor head into the used ring for use by the guest.

pub fn avail_idx<M: GuestMemory>(&self, mem: &M) -> Wrapping<u16>

Fetch the available ring index (virtq_avail->idx) from guest memory. This is written by the driver, to indicate the next slot that will be filled in the avail ring.

pub fn used_event<M: GuestMemory>(&self, mem: &M) -> Wrapping<u16>

Get the value of the used event field of the avail ring.

fn set_avail_event<M: GuestMemory>(&mut self, val: u16, mem: &M)

Helper method that writes val to the avail_event field of the used ring.

pub fn try_enable_notification<M: GuestMemory>(&mut self, mem: &M) -> bool

Try to enable notification events from the guest driver. Returns true if notifications were successfully enabled. Otherwise it means that one or more descriptors can still be consumed from the available ring and we can’t guarantee that there will be a notification. In this case the caller might want to consume the mentioned descriptors and call this method again.

pub fn enable_notif_suppression(&mut self)

Enable notification suppression.

pub fn prepare_kick<M: GuestMemory>(&mut self, mem: &M) -> bool

Check if we need to kick the guest.

Please note this method has side effects: once it returns true, it considers the driver will actually be notified, and won’t return true again until the driver updates used_event and/or the notification conditions hold once more.

This is similar to the vring_need_event() method implemented by the Linux kernel.