Trait vmm::devices::virtio::vhost_user::VhostUserHandleBackend

pub trait VhostUserHandleBackend: Sized {

    // Provided methods
    fn from_stream(_sock: UnixStream, _max_queue_num: u64) -> Self { ... }
    fn set_hdr_flags(&self, _flags: VhostUserHeaderFlag) { ... }
    fn get_features(&self) -> Result<u64, Error> { ... }
    fn set_features(&self, _features: u64) -> Result<(), Error> { ... }
    fn set_owner(&self) -> Result<(), Error> { ... }
    fn set_mem_table(
        &self,
        _regions: &[VhostUserMemoryRegionInfo]
    ) -> Result<(), Error> { ... }
    fn set_vring_num(&self, _queue_index: usize, _num: u16) -> Result<(), Error> { ... }
    fn set_vring_addr(
        &self,
        _queue_index: usize,
        _config_data: &VringConfigData
    ) -> Result<(), Error> { ... }
    fn set_vring_base(
        &self,
        _queue_index: usize,
        _base: u16
    ) -> Result<(), Error> { ... }
    fn set_vring_call(
        &self,
        _queue_index: usize,
        _fd: &EventFd
    ) -> Result<(), Error> { ... }
    fn set_vring_kick(
        &self,
        _queue_index: usize,
        _fd: &EventFd
    ) -> Result<(), Error> { ... }
    fn get_protocol_features(
        &mut self
    ) -> Result<VhostUserProtocolFeatures, Error> { ... }
    fn set_protocol_features(
        &mut self,
        _features: VhostUserProtocolFeatures
    ) -> Result<(), Error> { ... }
    fn set_vring_enable(
        &mut self,
        _queue_index: usize,
        _enable: bool
    ) -> Result<(), Error> { ... }
    fn get_config(
        &mut self,
        _offset: u32,
        _size: u32,
        _flags: VhostUserConfigFlags,
        _buf: &[u8]
    ) -> Result<(VhostUserConfig, VhostUserConfigPayload), Error> { ... }
    fn set_config(
        &mut self,
        _offset: u32,
        _flags: VhostUserConfigFlags,
        _buf: &[u8]
    ) -> Result<(), Error> { ... }

}

Provided Methods

fn from_stream(_sock: UnixStream, _max_queue_num: u64) -> Self

Constructor of Frontend

fn set_hdr_flags(&self, _flags: VhostUserHeaderFlag)

fn get_features(&self) -> Result<u64, Error>

Get from the underlying vhost implementation the feature bitmask.

fn set_features(&self, _features: u64) -> Result<(), Error>

Enable features in the underlying vhost implementation using a bitmask.

fn set_owner(&self) -> Result<(), Error>

Set the current Frontend as an owner of the session.

fn set_mem_table( &self, _regions: &[VhostUserMemoryRegionInfo] ) -> Result<(), Error>

Set the memory map regions on the slave so it can translate the vring addresses. In the ancillary data there is an array of file descriptors

fn set_vring_num(&self, _queue_index: usize, _num: u16) -> Result<(), Error>

Set the size of the queue.

fn set_vring_addr( &self, _queue_index: usize, _config_data: &VringConfigData ) -> Result<(), Error>

Sets the addresses of the different aspects of the vring.

fn set_vring_base(&self, _queue_index: usize, _base: u16) -> Result<(), Error>

Sets the base offset in the available vring.

fn set_vring_call( &self, _queue_index: usize, _fd: &EventFd ) -> Result<(), Error>

Set the event file descriptor to signal when buffers are used. Bits (0-7) of the payload contain the vring index. Bit 8 is the invalid FD flag. This flag is set when there is no file descriptor in the ancillary data. This signals that polling will be used instead of waiting for the call.

fn set_vring_kick( &self, _queue_index: usize, _fd: &EventFd ) -> Result<(), Error>

Set the event file descriptor for adding buffers to the vring. Bits (0-7) of the payload contain the vring index. Bit 8 is the invalid FD flag. This flag is set when there is no file descriptor in the ancillary data. This signals that polling should be used instead of waiting for a kick.

fn get_protocol_features(&mut self) -> Result<VhostUserProtocolFeatures, Error>

fn set_protocol_features( &mut self, _features: VhostUserProtocolFeatures ) -> Result<(), Error>

fn set_vring_enable( &mut self, _queue_index: usize, _enable: bool ) -> Result<(), Error>

fn get_config( &mut self, _offset: u32, _size: u32, _flags: VhostUserConfigFlags, _buf: &[u8] ) -> Result<(VhostUserConfig, VhostUserConfigPayload), Error>

fn set_config( &mut self, _offset: u32, _flags: VhostUserConfigFlags, _buf: &[u8] ) -> Result<(), Error>