Module vmm::devices::virtio::virtio_block::device

Macros

unwrap_async_file_engine_or_return 🔒

Structs

DiskProperties

Helper object for setting up all Block fields derived from its backing file.

VirtioBlock

Virtio device for exposing block level read/write operations on a host file.

VirtioBlockConfig

Use this structure to set up the Block Device before booting the kernel.

Enums

FileEngineType

The engine file type, either Sync or Async (through io_uring).