Module vmm::devices::virtio::virtio_block::metrics

Defines the metrics system for block devices.

Metrics format

The metrics are flushed in JSON when requested by vmm::logger::metrics::METRICS.write().

JSON example with metrics:

{
 "block_drv0": {
    "activate_fails": "SharedIncMetric",
    "cfg_fails": "SharedIncMetric",
    "no_avail_buffer": "SharedIncMetric",
    "event_fails": "SharedIncMetric",
    "execute_fails": "SharedIncMetric",
    ...
 }
 "block_drv1": {
    "activate_fails": "SharedIncMetric",
    "cfg_fails": "SharedIncMetric",
    "no_avail_buffer": "SharedIncMetric",
    "event_fails": "SharedIncMetric",
    "execute_fails": "SharedIncMetric",
    ...
 }
 ...
 "block_drive_id": {
    "activate_fails": "SharedIncMetric",
    "cfg_fails": "SharedIncMetric",
    "no_avail_buffer": "SharedIncMetric",
    "event_fails": "SharedIncMetric",
    "execute_fails": "SharedIncMetric",
    ...
 }
 "block": {
    "activate_fails": "SharedIncMetric",
    "cfg_fails": "SharedIncMetric",
    "no_avail_buffer": "SharedIncMetric",
    "event_fails": "SharedIncMetric",
    "execute_fails": "SharedIncMetric",
    ...
 }
}

Each block field in the example above is a serializable BlockDeviceMetrics structure collecting metrics such as activate_fails, cfg_fails, etc. for the block device. block_drv0 represent metrics for the endpoint “/drives/drv0”, block_drv1 represent metrics for the endpoint “/drives/drv1”, and block_drive_id represent metrics for the endpoint “/drives/{drive_id}” block device respectively and block is the aggregate of all the per device metrics.

Limitations

block device currently do not have vmm::logger::metrics::StoreMetrics so aggregate doesn’t consider them.

Design

The main design goals of this system are:

To improve block device metrics by logging them at per device granularity.

Continue to provide aggregate block metrics to maintain backward compatibility.

Move BlockDeviceMetrics out of from logger and decouple it.

Rely on serde to provide the actual serialization for writing the metrics.

Since all metrics start at 0, we implement the Default trait via derive for all of them, to avoid having to initialize everything by hand.

Devices could be created in any order i.e. the first device created could either be drv0 or drv1 so if we use a vector for BlockDeviceMetrics and call 1st device as block0, then block0 could sometimes point to drv0 and sometimes to drv1 which doesn’t help with analysing the metrics. So, use Map instead of Vec to help understand which drive the metrics actually belongs to.

The system implements 1 type of metrics:

Shared Incremental Metrics (SharedIncMetrics) - dedicated for the metrics which need a counter (i.e the number of times an API request failed). These metrics are reset upon flush. We add BlockDeviceMetrics entries from block::metrics::METRICS into Block device instead of Block device having individual separate BlockDeviceMetrics entries because Block device is not accessible from signal handlers to flush metrics and block::metrics::METRICS is.

Structs

BlockDeviceMetrics

Block Device associated metrics.

BlockMetricsPerDevice

map of block drive id and metrics this should be protected by a lock before accessing.

Statics

METRICS 🔒

Pool of block-related metrics per device behind a lock to keep things thread safe. Since the lock is initialized here it is safe to unwrap it without any check.

Functions

flush_metrics

This function facilitates aggregation and serialization of per block device metrics.