Struct vmm::devices::virtio::virtio_block::metrics::BlockDeviceMetrics

pub struct BlockDeviceMetrics {

    pub activate_fails: SharedIncMetric,
    pub cfg_fails: SharedIncMetric,
    pub no_avail_buffer: SharedIncMetric,
    pub event_fails: SharedIncMetric,
    pub execute_fails: SharedIncMetric,
    pub invalid_reqs_count: SharedIncMetric,
    pub flush_count: SharedIncMetric,
    pub queue_event_count: SharedIncMetric,
    pub rate_limiter_event_count: SharedIncMetric,
    pub update_count: SharedIncMetric,
    pub update_fails: SharedIncMetric,
    pub read_bytes: SharedIncMetric,
    pub write_bytes: SharedIncMetric,
    pub read_count: SharedIncMetric,
    pub write_count: SharedIncMetric,
    pub rate_limiter_throttled_events: SharedIncMetric,
    pub io_engine_throttled_events: SharedIncMetric,

}

Block Device associated metrics.

Fields
activate_fails: SharedIncMetric

Number of times when activate failed on a block device.

cfg_fails: SharedIncMetric

Number of times when interacting with the space config of a block device failed.

no_avail_buffer: SharedIncMetric

No available buffer for the block queue.

event_fails: SharedIncMetric

Number of times when handling events on a block device failed.

execute_fails: SharedIncMetric

Number of failures in executing a request on a block device.

invalid_reqs_count: SharedIncMetric

Number of invalid requests received for this block device.

flush_count: SharedIncMetric

Number of flushes operation triggered on this block device.

queue_event_count: SharedIncMetric

Number of events triggered on the queue of this block device.

rate_limiter_event_count: SharedIncMetric

Number of events ratelimiter-related.

update_count: SharedIncMetric

Number of update operation triggered on this block device.

update_fails: SharedIncMetric

Number of failures while doing update on this block device.

read_bytes: SharedIncMetric

Number of bytes read by this block device.

write_bytes: SharedIncMetric

Number of bytes written by this block device.

read_count: SharedIncMetric

Number of successful read operations.

write_count: SharedIncMetric

Number of successful write operations.

rate_limiter_throttled_events: SharedIncMetric

Number of rate limiter throttling events.

io_engine_throttled_events: SharedIncMetric

Number of virtio events throttled because of the IO engine. This happens when the io_uring submission queue is full.

Implementations

impl BlockDeviceMetrics

pub fn aggregate(&mut self, other: &Self)

block metrics are SharedIncMetric where the diff of current vs old is serialized i.e. serialize_u64(current-old). So to have the aggregate serialized in same way we need to fetch the diff of current vs old metrics and add it to the aggregate.