Module vmm::devices::virtio::vsock::csm::connection

This module implements our vsock connection state machine. The heavy lifting is done by connection::VsockConnection, while this file only defines some constants and helper structs.

Structs

VsockConnection

A self-managing connection object, that handles communication between a guest-side AF_VSOCK socket and a host-side ReadVolatile + Write + WriteVolatile + AsRawFd stream.

Traits

VsockConnectionBackend

Trait that vsock connection backends need to implement.