Struct vmm::devices::virtio::vsock::csm::connection::VsockConnection

pub struct VsockConnection<S: VsockConnectionBackend> {

    state: ConnState,
    local_cid: u64,
    peer_cid: u64,
    local_port: u32,
    peer_port: u32,
    stream: S,
    tx_buf: TxBuf,
    fwd_cnt: Wrapping<u32>,
    peer_buf_alloc: u32,
    peer_fwd_cnt: Wrapping<u32>,
    rx_cnt: Wrapping<u32>,
    last_fwd_cnt_to_peer: Wrapping<u32>,
    pending_rx: PendingRxSet,
    expiry: Option<Instant>,

}

A self-managing connection object, that handles communication between a guest-side AF_VSOCK socket and a host-side ReadVolatile + Write + WriteVolatile + AsRawFd stream.

Fields
state: ConnState

The current connection state.

local_cid: u64

The local CID. Most of the time this will be the constant 2 (the vsock host CID).

peer_cid: u64

The peer (guest) CID.

local_port: u32

The local (host) port.

peer_port: u32

The peer (guest) port.

stream: S

The (connected) host-side stream.

tx_buf: TxBuf

The TX buffer for this connection.

fwd_cnt: Wrapping<u32>

Total number of bytes that have been successfully written to self.stream, either directly, or flushed from self.tx_buf.

peer_buf_alloc: u32

The amount of buffer space that the peer (guest) has allocated for this connection.

peer_fwd_cnt: Wrapping<u32>

The total number of bytes that the peer has forwarded away.

rx_cnt: Wrapping<u32>

The total number of bytes sent to the peer (guest vsock driver)

last_fwd_cnt_to_peer: Wrapping<u32>

Our self.fwd_cnt, as last sent to the peer. This is used to provide proactive credit updates, and let the peer know it’s OK to send more data.

pending_rx: PendingRxSet

The set of pending RX packet indications that recv_pkt() will use to fill in a packet for the peer (guest).

expiry: Option<Instant>

Instant when this connection should be scheduled for immediate termination, due to some timeout condition having been fulfilled.

Implementations

impl<S> VsockConnection<S>
where S: VsockConnectionBackend + Debug,

pub fn new_peer_init( stream: S, local_cid: u64, peer_cid: u64, local_port: u32, peer_port: u32, peer_buf_alloc: u32 ) -> Self

Create a new guest-initiated connection object.

pub fn new_local_init( stream: S, local_cid: u64, peer_cid: u64, local_port: u32, peer_port: u32 ) -> Self

Create a new host-initiated connection object.

pub fn will_expire(&self) -> bool

Check if there is an expiry (kill) timer set for this connection, sometime in the future.

pub fn has_expired(&self) -> bool

Check if this connection needs to be scheduled for forceful termination, due to its kill timer having expired.

pub fn expiry(&self) -> Option<Instant>

Get the kill timer value, if one is set.

pub fn kill(&mut self)

Schedule the connection to be forcefully terminated ASAP (i.e. the next time the connection is asked to yield a packet, via recv_pkt()).

pub fn state(&self) -> ConnState

Return the connections state.

pub fn send_bytes_raw(&mut self, buf: &[u8]) -> Result<usize, VsockCsmError>

Send some raw, untracked, data straight to the underlying connected stream. Returns: number of bytes written, or the error describing the write failure.

Warning: this will bypass the connection state machine and write directly to the underlying stream. No account of this write is kept, which includes bypassing vsock flow control.

fn send_bytes(&mut self, pkt: &VsockPacket) -> Result<(), VsockError>

Send some raw data (a byte-slice) to the host stream.

Raw data can either be sent straight to the host stream, or to our TX buffer, if the former fails.

fn peer_needs_credit_update(&self) -> bool

Check if the credit information the peer has last received from us is outdated.

fn need_credit_update_from_peer(&self) -> bool

Check if we need to ask the peer for a credit update before sending any more data its way.

fn peer_avail_credit(&self) -> usize

Get the maximum number of bytes that we can send to our peer, without overflowing its buffer.

fn init_pkt<'a>(&self, pkt: &'a mut VsockPacket) -> &'a mut VsockPacket

Prepare a packet header for transmission to our peer.