Struct vmm::devices::virtio::vsock::packet::VsockPacket

pub struct VsockPacket {
    hdr: VsockPacketHeader,
    buffer: VsockPacketBuffer,
}

Struct describing a single vsock packet.

Encapsulates the virtio descriptor chain containing the packet through the IoVecBuffer[Mut] abstractions.

Fields
hdr: VsockPacketHeader

A copy of the vsock packet’s 44-byte header, held in hypervisor memory to minimize the number of accesses to guest memory. Can be written back to geust memory using VsockPacket::commit_hdr (only for RX buffers).

buffer: VsockPacketBuffer

The raw buffer, as it is contained in guest memory (containing both header and payload)

Implementations

impl VsockPacket

pub fn from_tx_virtq_head( chain: DescriptorChain<'_> ) -> Result<Self, VsockError>

Create the packet wrapper from a TX virtq chain head.

Errors

Returns

VsockError::UnreadableDescriptor if the provided descriptor chain contains any descriptor not marked as writable.

VsockError::DescChainTooShortForHeader if the descriptor chain’s total buffer length is insufficient to hold the 44 byte vsock header

VsockError::InvalidPktLen if the contained vsock header describes a vsock packet whose length would exceed [defs::MAX_PKT_BUR_SIZE].

VsockError::DescChainTooShortForPacket if the contained vsock header describes a vsock packet whose length exceeds the descriptor chain’s actual total buffer length.

pub fn from_rx_virtq_head( chain: DescriptorChain<'_> ) -> Result<Self, VsockError>

Create the packet wrapper from an RX virtq chain head.

Errors

Returns VsockError::DescChainTooShortForHeader if the descriptor chain’s total buffer length is insufficient to hold the 44 byte vsock header

pub fn hdr(&self) -> &VsockPacketHeader

Provides in-place access to the local copy of the vsock packet header.

pub fn commit_hdr(&mut self) -> Result<(), VsockError>

Writes the local copy of the packet header to the guest memory.

Errors

The function returns VsockError::UnwritableDescriptor if this VsockPacket contains a guest-to-host (TX) packet. It returned VsockError::InvalidPktLen if the packet’s payload as described by this VsockPacket would exceed defs::MAX_PKT_BUF_SIZE.

pub fn buf_size(&self) -> usize

Returns the total length of this VsockPacket’s buffer (e.g. the amount of data bytes contained in this packet).

Return value will equal the total length of the underlying descriptor chain’s buffers, minus the length of the vsock header.

pub fn read_at_offset_from<T: ReadVolatile + Debug>( &mut self, src: &mut T, offset: usize, count: usize ) -> Result<usize, VsockError>

pub fn write_from_offset_to<T: WriteVolatile + Debug>( &self, dst: &mut T, offset: usize, count: usize ) -> Result<usize, VsockError>

pub fn src_cid(&self) -> u64

pub fn set_src_cid(&mut self, cid: u64) -> &mut Self

pub fn dst_cid(&self) -> u64

pub fn set_dst_cid(&mut self, cid: u64) -> &mut Self

pub fn src_port(&self) -> u32

pub fn set_src_port(&mut self, port: u32) -> &mut Self

pub fn dst_port(&self) -> u32

pub fn set_dst_port(&mut self, port: u32) -> &mut Self

pub fn len(&self) -> u32

pub fn set_len(&mut self, len: u32) -> &mut Self

pub fn type_(&self) -> u16

pub fn set_type(&mut self, type_: u16) -> &mut Self

pub fn op(&self) -> u16

pub fn set_op(&mut self, op: u16) -> &mut Self

pub fn flags(&self) -> u32

pub fn set_flags(&mut self, flags: u32) -> &mut Self

pub fn set_flag(&mut self, flag: u32) -> &mut Self

pub fn buf_alloc(&self) -> u32

pub fn set_buf_alloc(&mut self, buf_alloc: u32) -> &mut Self

pub fn fwd_cnt(&self) -> u32

pub fn set_fwd_cnt(&mut self, fwd_cnt: u32) -> &mut Self