Struct vmm::devices::virtio::vsock::Vsock

pub struct Vsock<B> {
    cid: u64,
    pub(crate) queues: Vec<Queue>,
    pub(crate) queue_events: Vec<EventFd>,
    pub(crate) backend: B,
    pub(crate) avail_features: u64,
    pub(crate) acked_features: u64,
    pub(crate) irq_trigger: IrqTrigger,
    pub(crate) activate_evt: EventFd,
    pub(crate) device_state: DeviceState,
}

Structure representing the vsock device.

Fields
cid: u64queues: Vec<Queue>queue_events: Vec<EventFd>backend: Bavail_features: u64acked_features: u64irq_trigger: IrqTriggeractivate_evt: EventFddevice_state: DeviceState
Implementations

impl<B> Vsock<B>
where B: VsockBackend + Debug,

pub fn with_queues( cid: u64, backend: B, queues: Vec<VirtQueue> ) -> Result<Vsock<B>, VsockError>

Auxiliary function for creating a new virtio-vsock device with the given VM CID, vsock backend and empty virtio queues.

pub fn new(cid: u64, backend: B) -> Result<Vsock<B>, VsockError>

Create a new virtio-vsock device with the given VM CID and vsock backend.

pub fn id(&self) -> &str

Provides the ID of this vsock device as used in MMIO device identification.

pub fn cid(&self) -> u64

Retrieve the cid associated with this vsock device.

pub fn backend(&self) -> &B

Access the backend behind the device.

pub fn signal_used_queue(&self) -> Result<(), DeviceError>

Signal the guest driver that we’ve used some virtio buffers that it had previously made available.

pub fn process_rx(&mut self) -> bool

Walk the driver-provided RX queue buffers and attempt to fill them up with any data that we have pending. Return true if descriptors have been added to the used ring, and false otherwise.

pub fn process_tx(&mut self) -> bool

Walk the driver-provided TX queue buffers, package them up as vsock packets, and send them to the backend for processing. Return true if descriptors have been added to the used ring, and false otherwise.

pub fn send_transport_reset_event(&mut self) -> Result<(), DeviceError>

impl<B> Vsock<B>
where B: Debug + VsockBackend + 'static,

pub fn handle_rxq_event(&mut self, evset: EventSet) -> bool

pub fn handle_txq_event(&mut self, evset: EventSet) -> bool

pub fn handle_evq_event(&mut self, evset: EventSet) -> bool

pub fn notify_backend(&mut self, evset: EventSet) -> bool

Notify backend of new events.

fn register_runtime_events(&self, ops: &mut EventOps<'_>)

fn register_activate_event(&self, ops: &mut EventOps<'_>)

fn handle_activate_event(&self, ops: &mut EventOps<'_>)