Struct vmm::devices::virtio::vsock::VsockUnixBackend

pub struct VsockUnixBackend {
    cid: u64,
    conn_map: HashMap<ConnMapKey, VsockConnection<UnixStream>>,
    listener_map: HashMap<RawFd, EpollListener>,
    rxq: MuxerRxQ,
    killq: MuxerKillQ,
    host_sock: UnixListener,
    pub(crate) host_sock_path: String,
    epoll: Epoll,
    local_port_set: HashSet<u32>,
    local_port_last: u32,
}

The vsock connection multiplexer.

Fields
cid: u64

Guest CID.

conn_map: HashMap<ConnMapKey, VsockConnection<UnixStream>>

A hash map used to store the active connections.

listener_map: HashMap<RawFd, EpollListener>

A hash map used to store epoll event listeners / handlers.

rxq: MuxerRxQ

The RX queue. Items in this queue are consumed by VsockMuxer::recv_pkt(), and produced

by VsockMuxer::send_pkt() (e.g. RST in response to a connection request packet); and

in response to EPOLLIN events (e.g. data available to be read from an AF_UNIX socket).

killq: MuxerKillQ

A queue used for terminating connections that are taking too long to shut down.

host_sock: UnixListener

The Unix socket, through which host-initiated connections are accepted.

host_sock_path: String

The file system path of the host-side Unix socket. This is used to figure out the path to Unix sockets listening on specific ports. I.e. "<this path>_<port number>".

epoll: Epoll

The nested epoll event set, used to register epoll listeners.

local_port_set: HashSet<u32>

A hash set used to keep track of used host-side (local) ports, in order to assign local ports to host-initiated connections.

local_port_last: u32

The last used host-side port.

Implementations

impl VsockMuxer

pub fn new( cid: u64, host_sock_path: String ) -> Result<Self, VsockUnixBackendError>

Muxer constructor.

pub fn host_sock_path(&self) -> &str

Return the file system path of the host-side Unix socket.

fn handle_event(&mut self, fd: RawFd, event_set: EventSet)

Handle/dispatch an epoll event to its listener.

fn read_local_stream_port( stream: &mut UnixStream ) -> Result<u32, VsockUnixBackendError>

Parse a host “connect” command, and extract the destination vsock port.

fn add_connection( &mut self, key: ConnMapKey, conn: VsockConnection<UnixStream> ) -> Result<(), VsockUnixBackendError>

Add a new connection to the active connection pool.

fn remove_connection(&mut self, key: ConnMapKey)

Remove a connection from the active connection poll.

fn kill_connection(&mut self, key: ConnMapKey)

Schedule a connection for immediate termination. I.e. as soon as we can also let our peer know we’re dropping the connection, by sending it an RST packet.

fn add_listener( &mut self, fd: RawFd, listener: EpollListener ) -> Result<(), VsockUnixBackendError>

Register a new epoll listener under the muxer’s nested epoll FD.

fn remove_listener(&mut self, fd: RawFd) -> Option<EpollListener>

Remove (and return) a previously registered epoll listener.

fn allocate_local_port(&mut self) -> u32

Allocate a host-side port to be assigned to a new host-initiated connection.

fn free_local_port(&mut self, port: u32)

Mark a previously used host-side port as free.

fn handle_peer_request_pkt(&mut self, pkt: &VsockPacket)

Handle a new connection request comming from our peer (the guest vsock driver).

This will attempt to connect to a host-side Unix socket, expected to be listening at the file system path corresponing to the destination port. If successful, a new connection object will be created and added to the connection pool. On failure, a new RST packet will be scheduled for delivery to the guest.

fn apply_conn_mutation<F>(&mut self, key: ConnMapKey, mut_fn: F)
where F: FnOnce(&mut VsockConnection<UnixStream>),

Perform an action that might mutate a connection’s state.

This is used as shorthand for repetitive tasks that need to be performed after a connection object mutates. E.g.

update the connection’s epoll listener;

schedule the connection to be queried for RX data;

kill the connection if an unrecoverable error occurs.

fn sweep_killq(&mut self)

Check if any connections have timed out, and if so, schedule them for immediate termination.

fn enq_rst(&mut self, local_port: u32, peer_port: u32)

Enqueue an RST packet into self.rxq.

Enqueue errors aren’t propagated up the call chain, since there is nothing we can do to handle them. We do, however, log a warning, since not being able to enqueue an RST packet means we have to drop it, which is not normal operation.