Type Alias vmm::devices::virtio::vsock::unix::MuxerConnection

type MuxerConnection = VsockConnection<UnixStream>;

Aliased Type

struct MuxerConnection {

    state: ConnState,
    local_cid: u64,
    peer_cid: u64,
    local_port: u32,
    peer_port: u32,
    stream: UnixStream,
    tx_buf: TxBuf,
    fwd_cnt: Wrapping<u32>,
    peer_buf_alloc: u32,
    peer_fwd_cnt: Wrapping<u32>,
    rx_cnt: Wrapping<u32>,
    last_fwd_cnt_to_peer: Wrapping<u32>,
    pending_rx: PendingRxSet,
    expiry: Option<Instant>,

}

Fields
state: ConnState

The current connection state.

local_cid: u64

The local CID. Most of the time this will be the constant 2 (the vsock host CID).

peer_cid: u64

The peer (guest) CID.

local_port: u32

The local (host) port.

peer_port: u32

The peer (guest) port.

stream: UnixStream

The (connected) host-side stream.

tx_buf: TxBuf

The TX buffer for this connection.

fwd_cnt: Wrapping<u32>

Total number of bytes that have been successfully written to self.stream, either directly, or flushed from self.tx_buf.

peer_buf_alloc: u32

The amount of buffer space that the peer (guest) has allocated for this connection.

peer_fwd_cnt: Wrapping<u32>

The total number of bytes that the peer has forwarded away.

rx_cnt: Wrapping<u32>

The total number of bytes sent to the peer (guest vsock driver)

last_fwd_cnt_to_peer: Wrapping<u32>

Our self.fwd_cnt, as last sent to the peer. This is used to provide proactive credit updates, and let the peer know it’s OK to send more data.

pending_rx: PendingRxSet

The set of pending RX packet indications that recv_pkt() will use to fill in a packet for the peer (guest).

expiry: Option<Instant>

Instant when this connection should be scheduled for immediate termination, due to some timeout condition having been fulfilled.