Module vmm::dumbo

minimalist HTTP/TCP/IPv4 stack named DUMBO Provides helper logic for parsing and writing protocol data units, and minimalist implementations of a TCP listener, a TCP connection, and an HTTP/1.1 server.

Re-exports

pub use crate::dumbo::pdu::arp::EthIPv4ArpFrame;

pub use crate::dumbo::pdu::arp::ETH_IPV4_FRAME_LEN;

pub use crate::dumbo::pdu::ethernet::EthernetFrame;

pub use crate::dumbo::pdu::ethernet::ETHERTYPE_ARP;

pub use crate::dumbo::pdu::ethernet::ETHERTYPE_IPV4;

pub use crate::dumbo::pdu::ethernet::PAYLOAD_OFFSET as ETHERNET_PAYLOAD_OFFSET;

pub use crate::dumbo::pdu::ipv4::IPv4Packet;

pub use crate::dumbo::pdu::ipv4::PROTOCOL_TCP;

pub use crate::dumbo::pdu::ipv4::PROTOCOL_UDP;

pub use crate::dumbo::pdu::udp::UdpDatagram;

pub use crate::dumbo::pdu::udp::UDP_HEADER_SIZE;

Modules

pdu

A module for interpreting byte slices as protocol data units (PDUs).

tcp

Provides functionality for handling incoming TCP connections.

Traits

ByteBuffer

Represents a generalization of a borrowed [u8] slice.