Module vmm::dumbo::pdu

A module for interpreting byte slices as protocol data units (PDUs).

A PDU represents data transmitted as a single unit during communication using a specific protocol. Ethernet frames, IP packets, and TCP segments are all examples of protocol data units.

Modules

arp

Contains logic that helps with handling ARP frames over Ethernet, which encapsulate requests or replies related to IPv4 addresses.

bytes

Defines traits which allow byte slices to be interpreted as sequences of bytes that stand for different values packed together using network byte ordering (such as network packets).

ethernet

Contains support for parsing and writing Ethernet frames. Does not currently offer support for 802.1Q tags.

ipv4

Contains support for parsing and writing IPv4 packets.

tcp

Contains support for parsing and writing TCP segments.

udp

Contains support for parsing and writing User Datagram Protocol (UDP) packets, with no support for jumbograms.

Structs

Incomplete

This is the baseline definition of the Incomplete struct, which wraps a PDU that does is still missing some values or content.

Enums

ChecksumProto 🔒

Functions

compute_checksum 🔒

Computes the checksum of a TCP/UDP packet. Since both protocols use the same algorithm to compute the checksum.