Module vmm::dumbo::pdu::arp

Contains logic that helps with handling ARP frames over Ethernet, which encapsulate requests or replies related to IPv4 addresses.

A more detailed view of an ARP frame can be found here.

Structs

EthIPv4ArpFrame

The inner bytes will be interpreted as an ARP frame.

Enums

Error

Represents errors which may occur while parsing or writing a frame.

Constants

ETH_IPV4_FRAME_LEN

The length of an ARP frame for IPv4 over Ethernet.

ETH_IPV4_SPA_OFFSET 🔒

ETH_IPV4_THA_OFFSET 🔒

ETH_IPV4_TPA_OFFSET 🔒

HLEN_OFFSET 🔒

HTYPE_ETHERNET

ARP is for Ethernet hardware

HTYPE_OFFSET 🔒

IPV4_ADDR_LEN 🔒

OPER_OFFSET 🔒

OPER_REPLY

ARP Reply operation

OPER_REQUEST

ARP Request operation

PLEN_OFFSET 🔒

PTYPE_OFFSET 🔒

SHA_OFFSET 🔒

_DERIVE_Display_FOR_Error 🔒

Functions

test_speculative_tpa

This function checks if buf may hold an Ethernet frame which encapsulates an EthIPv4ArpRequest for the given address. Cannot produce false negatives.