Struct vmm::dumbo::pdu::arp::EthIPv4ArpFrame

pub struct EthIPv4ArpFrame<'a, T: 'a> { /* private fields */ }

The inner bytes will be interpreted as an ARP frame.

ARP is a generic protocol as far as data link layer and network layer protocols go, but this particular implementation is concerned with ARP frames related to IPv4 over Ethernet.

Implementations

impl<'a, T: NetworkBytes + Debug> EthIPv4ArpFrame<'a, T>

pub fn from_bytes_unchecked(bytes: T) -> Self

Interprets the given bytes as an ARP frame, without doing any validity checks beforehand.

Panics

This method does not panic, but further method calls on the resulting object may panic if bytes contains invalid input.

pub fn request_from_bytes(bytes: T) -> Result<Self, Error>

Tries to interpret a byte slice as a valid IPv4 over Ethernet ARP request.

If no error occurs, it guarantees accessor methods (which make use of various _unchecked functions) are safe to call on the result, because all predefined offsets will be valid.

pub fn htype(&self) -> u16

Returns the hardware type of the frame.

pub fn ptype(&self) -> u16

Returns the protocol type of the frame.

pub fn hlen(&self) -> u8

Returns the hardware address length of the frame.

pub fn plen(&self) -> u8

Returns the protocol address length of the frame.

pub fn operation(&self) -> u16

Returns the type of operation within the frame.

pub fn sha(&self) -> MacAddr

Returns the sender hardware address.

pub fn spa(&self) -> Ipv4Addr

Returns the sender protocol address.

pub fn tha(&self) -> MacAddr

Returns the target hardware address.

pub fn tpa(&self) -> Ipv4Addr

Returns the target protocol address.

pub fn len(&self) -> usize

Returns the length of the frame.

impl<'a, T: NetworkBytesMut + Debug> EthIPv4ArpFrame<'a, T>

pub fn write_request( buf: T, sha: MacAddr, spa: Ipv4Addr, tha: MacAddr, tpa: Ipv4Addr ) -> Result<Self, Error>

Attempts to write an ARP request to buf, based on the specified hardware and protocol addresses.

pub fn write_reply( buf: T, sha: MacAddr, spa: Ipv4Addr, tha: MacAddr, tpa: Ipv4Addr ) -> Result<Self, Error>

Attempts to write an ARP reply to buf, based on the specified hardware and protocol addresses.

pub fn set_htype(&mut self, value: u16)

Sets the hardware type of the frame.

pub fn set_ptype(&mut self, value: u16)

Sets the protocol type of the frame.

pub fn set_hlen(&mut self, value: u8)

Sets the hardware address length of the frame.

pub fn set_plen(&mut self, value: u8)

Sets the protocol address length of the frame.

pub fn set_operation(&mut self, value: u16)

Sets the operation within the frame.

pub fn set_sha(&mut self, addr: MacAddr)

Sets the sender hardware address.

pub fn set_spa(&mut self, addr: Ipv4Addr)

Sets the sender protocol address.

pub fn set_tha(&mut self, addr: MacAddr)

Sets the target hardware address.

pub fn set_tpa(&mut self, addr: Ipv4Addr)

Sets the target protocol address.