Module vmm::dumbo::pdu::ipv4

Contains support for parsing and writing IPv4 packets.

A picture of the IPv4 packet header can be found here (watch out for the MSB 0 bit numbering).

Structs

IPv4Packet

Interprets the inner bytes as an IPv4 packet.

Enums

Error

Describes the errors which may occur while handling IPv4 packets.

Constants

DEFAULT_TTL

Default TTL value

DESTINATION_ADDRESS_OFFSET 🔒

DSCP_AND_ECN_OFFSET 🔒

FLAGS_AND_FRAGMENTOFF_OFFSET 🔒

HEADER_CHECKSUM_OFFSET 🔒

IDENTIFICATION_OFFSET 🔒

IPV4_VERSION

Indicates version 4 of the IP protocol

OPTIONS_OFFSET 🔒

PROTOCOL_OFFSET 🔒

PROTOCOL_TCP

The IP protocol number associated with TCP.

PROTOCOL_UDP

The IP protocol number associated with UDP.

SOURCE_ADDRESS_OFFSET 🔒

TOTAL_LEN_OFFSET 🔒

TTL_OFFSET 🔒

VERSION_AND_IHL_OFFSET 🔒

_DERIVE_Display_FOR_Error 🔒

Functions

test_speculative_dst_addr

This function checks if buf may hold an IPv4Packet heading towards the given address. Cannot produce false negatives.