Struct vmm::dumbo::pdu::ipv4::IPv4Packet

pub struct IPv4Packet<'a, T: 'a> {
    bytes: InnerBytes<'a, T>,
}

Interprets the inner bytes as an IPv4 packet.

Fields
bytes: InnerBytes<'a, T>
Implementations

impl<'a, T: NetworkBytes + Debug> IPv4Packet<'a, T>

pub fn from_bytes_unchecked(bytes: T) -> Self

Interpret bytes as an IPv4Packet without checking the validity of the header fields, and the length of the inner byte sequence.

Panics

This method does not panic, but further method calls on the resulting object may panic if bytes contains invalid input.

pub fn from_bytes(bytes: T, verify_checksum: bool) -> Result<Self, Error>

Attempts to interpret bytes as an IPv4 packet, checking the validity of the header fields and the length of the inner byte sequence.

pub fn version_and_header_len(&self) -> (u8, u8)

Returns the value of the version header field, and the header length.

This method returns the actual length (in bytes) of the header, and not the value of the ihl header field).

pub fn header_len(&self) -> u8

Returns the packet header length (in bytes).

pub fn dscp_and_ecn(&self) -> (u8, u8)

Returns the values of the dscp and ecn header fields.

pub fn total_len(&self) -> u16

Returns the value of the ‘total length’ header field.

pub fn identification(&self) -> u16

Returns the value of the identification header field.

pub fn flags_and_fragment_offset(&self) -> (u8, u16)

Returns the values of the flags and fragment offset header fields.

pub fn ttl(&self) -> u8

Returns the value of the ttl header field.

pub fn protocol(&self) -> u8

Returns the value of the protocol header field.

pub fn header_checksum(&self) -> u16

Returns the value of the header checksum header field.

pub fn source_address(&self) -> Ipv4Addr

Returns the source IPv4 address of the packet.

pub fn destination_address(&self) -> Ipv4Addr

Returns the destination IPv4 address of the packet.

pub fn payload_unchecked(&self, header_len: usize) -> &[u8]

Returns a byte slice containing the payload, using the given header length value to compute the payload offset.

Panics

This method may panic if the value of header_len is invalid.

pub fn payload(&self) -> &[u8]

Returns a byte slice that contains the payload of the packet.

pub fn len(&self) -> usize

Returns the length of the inner byte sequence.

This is equal to the output of the total_len() method for properly constructed instances of IPv4Packet.

pub fn compute_checksum_unchecked(&self, header_len: usize) -> u16

Computes and returns the packet header checksum using the provided header length.

A nice description of how this works can be found here. May panic for invalid values of header_len.

Panics

This method may panic if the value of header_len is invalid.

pub fn compute_checksum(&self) -> u16

Computes and returns the packet header checksum.

impl<'a, T: NetworkBytesMut + Debug> IPv4Packet<'a, T>

pub fn write_header( buf: T, protocol: u8, src_addr: Ipv4Addr, dst_addr: Ipv4Addr ) -> Result<Incomplete<Self>, Error>

Attempts to write an IPv4 packet header to buf, making sure there is enough space.

This method returns an incomplete packet, because the size of the payload might be unknown at this point. IP options are not allowed, which means header_len == OPTIONS_OFFSET. The dscp, ecn, identification, flags, and fragment_offset fields are set to 0. The ttl is set to a default value. The total_len and checksum fields will be set when the length of the incomplete packet is determined.

pub fn set_version_and_header_len( &mut self, version: u8, header_len: u8 ) -> &mut Self

Sets the values of the version and ihl header fields (the latter is computed from the value of header_len).

pub fn set_dscp_and_ecn(&mut self, dscp: u8, ecn: u8) -> &mut Self

Sets the values of the dscp and ecn header fields.

pub fn set_total_len(&mut self, value: u16) -> &mut Self

Sets the value of the total length header field.

pub fn set_identification(&mut self, value: u16) -> &mut Self

Sets the value of the identification header field.

pub fn set_flags_and_fragment_offset( &mut self, flags: u8, fragment_offset: u16 ) -> &mut Self

Sets the values of the flags and fragment offset header fields.

pub fn set_ttl(&mut self, value: u8) -> &mut Self

Sets the value of the ttl header field.

pub fn set_protocol(&mut self, value: u8) -> &mut Self

Sets the value of the protocol header field.

pub fn set_header_checksum(&mut self, value: u16) -> &mut Self

Sets the value of the header checksum header field.

pub fn set_source_address(&mut self, addr: Ipv4Addr) -> &mut Self

Sets the source address of the packet.

pub fn set_destination_address(&mut self, addr: Ipv4Addr) -> &mut Self

Sets the destination address of the packet.

pub fn payload_mut_unchecked(&mut self, header_len: usize) -> &mut [u8]

Returns a mutable byte slice representing the payload of the packet, using the provided header length to compute the payload offset.

Panics

This method may panic if the value of header_len is invalid.

pub fn payload_mut(&mut self) -> &mut [u8]

Returns a mutable byte slice representing the payload of the packet.