Struct vmm::dumbo::pdu::tcp::TcpSegment

pub struct TcpSegment<'a, T: 'a> { /* private fields */ }

Interprets the inner bytes as a TCP segment.

Implementations

impl<'a, T: NetworkBytes + Debug> TcpSegment<'a, T>

pub fn source_port(&self) -> u16

Returns the source port.

pub fn destination_port(&self) -> u16

Returns the destination port.

pub fn sequence_number(&self) -> u32

Returns the sequence number.

pub fn ack_number(&self) -> u32

Returns the acknowledgement number (only valid if the ACK flag is set).

pub fn header_len_rsvd_ns(&self) -> (u8, u8, bool)

Returns the header length, the value of the reserved bits, and whether the NS flag is set or not.

pub fn header_len(&self) -> u8

Returns the length of the header.

pub fn flags_after_ns(&self) -> Flags

Returns the TCP header flags, with the exception of NS.

pub fn window_size(&self) -> u16

Returns the value of the window size header field.

pub fn checksum(&self) -> u16

Returns the value of the checksum header field.

pub fn urgent_pointer(&self) -> u16

Returns the value of the urgent pointer header field (only valid if the URG flag is set).

pub fn options_unchecked(&self, header_len: usize) -> &[u8]

Returns the TCP header options as an [&u8] slice.

Panics

This method may panic if the value of header_len is invalid.

pub fn payload_unchecked(&self, header_len: usize) -> &[u8]

Returns a slice which contains the payload of the segment. May panic if the value of header_len is invalid.

Panics

This method may panic if the value of header_len is invalid.

pub fn len(&self) -> u16

Returns the length of the segment.

pub fn payload(&self) -> &[u8]

Returns a slice which contains the payload of the segment.

pub fn payload_len(&self) -> u16

Returns the length of the payload.

pub fn compute_checksum(&self, src_addr: Ipv4Addr, dst_addr: Ipv4Addr) -> u16

Computes the TCP checksum of the segment. More details about TCP checksum computation can be found here.

pub fn parse_mss_option_unchecked( &self, header_len: usize ) -> Result<Option<NonZeroU16>, Error>

Parses TCP header options (only MSS is supported for now).

If no error is encountered, returns the MSS value, or None if the option is not present.

Panics

This method may panic if the value of header_len is invalid.

pub fn from_bytes_unchecked(bytes: T) -> Self

Interprets bytes as a TCP segment without any validity checks.

Panics

This method does not panic, but further method calls on the resulting object may panic if bytes contains invalid input.

pub fn from_bytes( bytes: T, verify_checksum: Option<(Ipv4Addr, Ipv4Addr)> ) -> Result<Self, Error>

Attempts to interpret bytes as a TCP segment, checking the validity of the header fields.

The verify_checksum parameter must contain the source and destination addresses from the enclosing IPv4 packet if the TCP checksum must be validated.

impl<'a, T: NetworkBytesMut + Debug> TcpSegment<'a, T>

pub fn set_source_port(&mut self, value: u16) -> &mut Self

Sets the source port.

pub fn set_destination_port(&mut self, value: u16) -> &mut Self

Sets the destination port.

pub fn set_sequence_number(&mut self, value: u32) -> &mut Self

Sets the value of the sequence number field.

pub fn set_ack_number(&mut self, value: u32) -> &mut Self

Sets the value of the acknowledgement number field.

pub fn set_header_len_rsvd_ns(&mut self, header_len: u8, ns: bool) -> &mut Self

Sets the value of the ihl header field based on header_len (which should be a multiple of 4), clears the reserved bits, and sets the NS flag according to the last parameter.

pub fn set_flags_after_ns(&mut self, flags: Flags) -> &mut Self

Sets the value of the header byte containing every TCP flag except NS.

pub fn set_window_size(&mut self, value: u16) -> &mut Self

Sets the value of the window size field.

pub fn set_checksum(&mut self, value: u16) -> &mut Self

Sets the value of the checksum field.

pub fn set_urgent_pointer(&mut self, value: u16) -> &mut Self

Sets the value of the urgent pointer field.

pub fn payload_mut_unchecked(&mut self, header_len: usize) -> &mut [u8]

Returns a mutable slice containing the segment payload.

Panics

This method may panic if the value of header_len is invalid.

pub fn payload_mut(&mut self) -> &mut [u8]

Returns a mutable slice containing the segment payload.

pub fn write_segment<R: ByteBuffer + ?Sized + Debug>( buf: T, src_port: u16, dst_port: u16, seq_number: u32, ack_number: u32, flags_after_ns: Flags, window_size: u16, mss_option: Option<u16>, mss_remaining: u16, payload: Option<(&R, usize)>, compute_checksum: Option<(Ipv4Addr, Ipv4Addr)> ) -> Result<Self, Error>

Writes a complete TCP segment.

Arguments

buf - Write the segment to this buffer.

src_port - Source port.

dst_port - Destination port.

seq_number - Sequence number.

ack_number - Acknowledgement number.

flags_after_ns - TCP flags to set (except NS, which is always set to 0).

window_size - Value to write in the window size field.

mss_option - When a value is specified, use it to add a TCP MSS option to the header.

mss_remaining - Represents an upper bound on the payload length (the number of bytes used up by things like IP options have to be subtracted from the MSS). There is some redundancy looking at this argument and the next one, so we might end up removing or changing something.

payload - May contain a buffer which holds payload data and the maximum amount of bytes we should read from that buffer. When None, the TCP segment will carry no payload.

compute_checksum - May contain the pair addresses from the enclosing IPv4 packet, which are required for TCP checksum computation. Skip the checksum altogether when None.

pub fn write_incomplete_segment<R: ByteBuffer + ?Sized + Debug>( buf: T, seq_number: u32, ack_number: u32, flags_after_ns: Flags, window_size: u16, mss_option: Option<u16>, mss_remaining: u16, payload: Option<(&R, usize)> ) -> Result<Incomplete<Self>, Error>

Writes an incomplete TCP segment, which is missing the source port, destination port, and checksum fields.

This method writes the rest of the segment, including data (when available). Only the MSS option is supported for now. The NS flag, URG flag, and urgent pointer field are set to 0.

Arguments

buf - Write the segment to this buffer.

seq_number - Sequence number.

ack_number - Acknowledgement number.

flags_after_ns - TCP flags to set (except NS, which is always set to 0).

window_size - Value to write in the window size field.

mss_option - When a value is specified, use it to add a TCP MSS option to the header.

mss_remaining - Represents an upper bound on the payload length (the number of bytes used up by things like IP options have to be subtracted from the MSS). There is some redundancy looking at this argument and the next one, so we might end up removing or changing something.

payload - May contain a buffer which holds payload data and the maximum amount of bytes we should read from that buffer. When None, the TCP segment will carry no payload.