Type Alias vmm::dumbo::tcp::connection::PayloadSource

pub type PayloadSource<'a, R> = Option<(&'a R, Wrapping<u32>)>;

Defines a segment payload source.

When not None, it contains a ByteBuffer which holds the actual data, and the sequence number associated with the first byte from the buffer.

Aliased Type

enum PayloadSource<'a, R> {
    None,
    Some((&'a R, Wrapping<u32>)),
}

Variants

None

No value.

Some((&'a R, Wrapping<u32>))

Some value of type T.