Module vmm::dumbo::tcp::handler

Exposes simple TCP over IPv4 listener functionality via the TcpIPv4Handler structure.

Structs

ConnectionTuple 🔒

TcpIPv4Handler

Implements a minimalist TCP over IPv4 listener.

Enums

RecvError

Describes errors which may be encountered by the receive_packet method from TcpIPv4Handler.

RecvEvent

Describes events which may occur when the handler receives packets.

RecvSegmentOutcome 🔒

WriteEvent

Describes events which may occur when the handler writes packets.

WriteNextError

Describes errors which may be encountered by the write_next_packet method from TcpIPv4Handler.

Constants

_DERIVE_Display_FOR_RecvError 🔒

_DERIVE_Display_FOR_WriteNextError 🔒