Struct vmm::io_uring::IoUring

pub struct IoUring {
    registered_fds_count: u32,
    squeue: SubmissionQueue,
    cqueue: CompletionQueue,
    fd: File,
    num_ops: u32,
}

Main object representing an io_uring instance.

Fields
registered_fds_count: u32squeue: SubmissionQueuecqueue: CompletionQueuefd: Filenum_ops: u32
Implementations

impl IoUring

pub fn new( num_entries: u32, files: Vec<&File>, restrictions: Vec<Restriction>, eventfd: Option<RawFd> ) -> Result<Self, IoUringError>

Create a new instance.

Arguments

num_entries - Requested number of entries in the ring. Will be rounded up to the nearest power of two.

files - Files to be registered for IO.

restrictions - Vector of Restrictions

eventfd - Optional eventfd for receiving completion notifications.

pub unsafe fn push<T: Debug>( &mut self, op: Operation<T> ) -> Result<(), (IoUringError, T)>

Push an Operation onto the submission queue.

Safety

Unsafe because we pass a raw user_data pointer to the kernel. It’s up to the caller to make sure that this value is ever freed (not leaked).

pub unsafe fn pop<T: Debug>(&mut self) -> Result<Option<Cqe<T>>, IoUringError>

Pop a completed entry off the completion queue. Returns Ok(None) if there are no entries. The type T must be the same as the user_data type used for push-ing the operation.

Safety

Unsafe because we reconstruct the user_data from a raw pointer passed by the kernel. It’s up to the caller to make sure that T is the correct type of the user_data, that the raw pointer is valid and that we have full ownership of that address.

fn do_submit(&mut self, min_complete: u32) -> Result<u32, IoUringError>

pub fn submit(&mut self) -> Result<u32, IoUringError>

Submit all operations but don’t wait for any completions.

pub fn submit_and_wait_all(&mut self) -> Result<u32, IoUringError>

Submit all operations and wait for their completion.

pub fn pending_sqes(&self) -> Result<u32, IoUringError>

Return the number of operations currently on the submission queue.

pub fn num_ops(&self) -> u32

A total of the number of ops in the submission and completion queues, as well as the in-flight ops.

fn enable(&mut self) -> Result<(), IoUringError>

fn register_files(&mut self, files: Vec<&File>) -> Result<(), IoUringError>

fn register_eventfd(&self, fd: RawFd) -> Result<(), IoUringError>

fn register_restrictions( &self, restrictions: Vec<Restriction> ) -> Result<(), IoUringError>

fn check_features(params: io_uring_params) -> Result<(), IoUringError>

fn check_operations(&self) -> Result<(), IoUringError>