Module vmm::logger::logging

Structs

LevelFilterFromStrError

Error type for [<LevelFilter as FromStr>::from_str].

LogFilter

LogFormat

Logger

LoggerConfig

Strongly typed structure used to describe the logger.

LoggerConfiguration

LoggerUpdateError

Error type for Logger::update.

Enums

LevelFilter

This is required since we originally supported Warning and uppercase variants being used as the log level filter. It would be a breaking change to no longer support this. In the next breaking release this should be removed (replaced with log::LevelFilter and only supporting its default deserialization).

Constants

DEFAULT_INSTANCE_ID

Default instance id.

DEFAULT_LEVEL

Default level filter for logger matching the swagger specification (src/api_server/swagger/firecracker.yaml).

Statics

INSTANCE_ID

Instance id.

LOGGER

The logger.

Type Aliases

LoggerInitError

Error type for Logger::init.