Module vmm::logger::metrics

Defines the metrics system.

Metrics format

The metrics are flushed in JSON format each 60 seconds. The first field will always be the timestamp followed by the JSON representation of the structures representing each component on which we are capturing specific metrics.

JSON example with metrics:

{
 "utc_timestamp_ms": 1541591155180,
 "api_server": {
   "process_startup_time_us": 0,
   "process_startup_time_cpu_us": 0
 },
 "block": {
   "activate_fails": 0,
   "cfg_fails": 0,
   "event_fails": 0,
   "flush_count": 0,
   "queue_event_count": 0,
   "read_count": 0,
   "write_count": 0
 }
}

The example above means that inside the structure representing all the metrics there is a field named block which is in turn a serializable child structure collecting metrics for the block device such as activate_fails, cfg_fails, etc.

Limitations

Metrics are only written to buffers.

Design

The main design goals of this system are:

Use lockless operations, preferably ones that don’t require anything other than simple reads/writes being atomic.

Exploit interior mutability and atomics being Sync to allow all methods (including the ones which are effectively mutable) to be callable on a global non-mut static.

Rely on serde to provide the actual serialization for writing the metrics.

Since all metrics start at 0, we implement the Default trait via derive for all of them, to avoid having to initialize everything by hand.

The system implements 2 types of metrics:

Shared Incremental Metrics (SharedIncMetrics) - dedicated for the metrics which need a counter (i.e the number of times an API request failed). These metrics are reset upon flush.

Shared Store Metrics (SharedStoreMetrics) - are targeted at keeping a persistent value, it is not intended to act as a counter (i.e for measure the process start up time for example).

The current approach for the SharedIncMetrics type is to store two values (current and previous) and compute the delta between them each time we do a flush (i.e by serialization). There are a number of advantages to this approach, including:

We don’t have to introduce an additional write (to reset the value) from the thread which does to actual writing, so less synchronization effort is required.

We don’t have to worry at all that much about losing some data if writing fails for a while (this could be a concern, I guess). If if turns out this approach is not really what we want, it’s pretty easy to resort to something else, while working behind the same interface.

Macros

create_serialize_proxy 🔒

Structs

ApiServerMetrics

Metrics related to the internal API server.

BalloonMetricsSerializeProxy

BlockMetricsSerializeProxy

DeprecatedApiMetrics

Metrics related to deprecated user-facing API calls.

EntropyMetricsSerializeProxy

FirecrackerMetrics

Structure storing all metrics while enforcing serialization support on them.

GetRequestsMetrics

Metrics specific to GET API Requests for counting user triggered actions and/or failures.

LatencyAggregateMetrics

Used to record Aggregate (min/max/sum) of latency metrics

LatencyMetricsRecorder

Provides efficient way to record LatencyAggregateMetrics

LegacyDevMetricsSerializeProxy

LoggerSystemMetrics

Metrics for the logging subsystem.

Metrics

Metrics system.

MmdsMetrics

Metrics for the MMDS functionality.

NetMetricsSerializeProxy

PatchRequestsMetrics

Metrics specific to PATCH API Requests for counting user triggered actions and/or failures.

PerformanceMetrics

Performance metrics related for the moment only to snapshots.

ProcessTimeReporter

Reporter object which computes the process wall time and process CPU time and populates the metric with the results.

PutRequestsMetrics

Metrics specific to PUT API Requests for counting user triggered actions and/or failures.

SeccompMetrics

Metrics for the seccomp filtering.

SerializeToUtcTimestampMs 🔒

SharedIncMetric

Representation of a metric that is expected to be incremented from more than one thread, so more synchronization is necessary.

SharedStoreMetric

Representation of a metric that is expected to hold a value that can be accessed from more than one thread, so more synchronization is necessary.

SignalMetrics

Metrics related to signals. Deadly signals must be of SharedStoreMetric type, since they can ever be either 0 or 1. This avoids a tricky race condition caused by the unatomic serialize method of SharedIncMetric, between two threads calling METRICS.write().

VcpuMetrics

Structure provides Metrics specific to VCPUs’ mode of functioning. Sample_count or number of kvm exits for IO and MMIO VM exits are covered by: exit_io_in, exit_io_out, exit_mmio_read and , exit_mmio_write. Count of other vm exits for events like shutdown/hlt/errors are covered by existing “failures” metric. The only vm exit for which sample_count is not covered is system event reset/shutdown but that should be fine since they are not failures and the vm is terminated anyways. LatencyAggregateMetrics only covers minimum, maximum and sum because average can be deduced from available metrics. e.g. dividing exit_io_in_agg.sum_us by exit_io_in` gives average of KVM exits handling input IO.

VhostUserMetricsSerializeProxy

VmmMetrics

Metrics specific to the machine manager as a whole.

VsockMetricsSerializeProxy

Enums

MetricsError

Describes the errors which may occur while handling metrics scenarios.

Constants

_DERIVE_Display_FOR_MetricsError 🔒

Statics

METRICS

Static instance used for handling metrics.

Traits

IncMetric

Used for defining new types of metrics that act as a counter (i.e they are continuously updated by incrementing their value).

StoreMetric

Used for defining new types of metrics that do not need a counter and act as a persistent indicator.