Struct vmm::mmds::data_store::Mmds

pub struct Mmds {
    data_store: Value,
    token_authority: Option<TokenAuthority>,
    is_initialized: bool,
    data_store_limit: usize,
}

The Mmds is the Microvm Metadata Service represented as an untyped json.

Fields
data_store: Valuetoken_authority: Option<TokenAuthority>is_initialized: booldata_store_limit: usize
Implementations

impl Mmds

pub fn default_with_limit(data_store_limit: usize) -> Self

MMDS default instance with limit data_store_limit

fn check_data_store_initialized(&self) -> Result<(), Error>

This method is needed to check if data store is initialized. When a PATCH request is made on an uninitialized Mmds structure this method should return a NotFound error.

pub fn set_version(&mut self, version: MmdsVersion) -> Result<(), Error>

Set the MMDS version.

pub fn version(&self) -> MmdsVersion

Return the MMDS version by checking the token authority field.

pub fn set_aad(&mut self, instance_id: &str)

Sets the Additional Authenticated Data to be used for encryption and decryption of the session token when MMDS version 2 is enabled.

pub fn is_valid_token(&self, token: &str) -> Result<bool, TokenError>

Checks if the provided token has not expired.

pub fn generate_token(&mut self, ttl_seconds: u32) -> Result<String, TokenError>

Generate a new Mmds token using the token authority.

pub fn set_data_store_limit(&mut self, data_store_limit: usize)

set MMDS data store limit to data_store_limit

pub fn put_data(&mut self, data: Value) -> Result<(), Error>

put data in MMDS data store

pub fn patch_data(&mut self, patch_data: Value) -> Result<(), Error>

patch update MMDS data store with patch_data

pub fn data_store_value(&self) -> Value

return MMDS data store value We do not check size of data_store before returning a result because due to limit from put/patch the data_store can not be bigger than the limit imposed by the server.

fn format_imds(json: &Value) -> Result<String, Error>

Returns the serde::Value in IMDS format plaintext. Currently, only JSON objects and strings can be IMDS formatted.

See the docs for detailed description of the IMDS format: https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-instance-metadata.html

Examples

{
    "key1" : {
        "key11": "value11"
        "key12": "value12"
    }
    "key2" : "value3"
    "key3" : "value3"
}

IMDS formatted JSON object:

key1/
key2
key3

JSON string:

"value"

IMDS formatted string:

value

If the serde_json::Value is not supported, an UnsupportedValueType error is returned.

pub fn get_value( &self, path: String, format: OutputFormat ) -> Result<String, Error>

Returns the subtree located at path. When the path corresponds to a leaf, it returns the value. Returns Error::NotFound when the path is invalid.