Module vmm::mmds::token

Structs

Token 🔒

Structure for token information.

TokenAuthority

Enums

Error

Constants

DESERIALIZATION_BYTES_LIMIT 🔒

Byte limit passed to bincode to guard against allocating too much memory when deserializing tokens.

IV_LEN

Length of initialization vector.

KEY_LEN

Length of the key used for encryption.

MAX_TOKEN_TTL_SECONDS

Maximum lifetime of token.

MILLISECONDS_PER_SECOND

Constant to convert seconds to milliseconds.

MIN_TOKEN_TTL_SECONDS

Minimum lifetime of token.

PATH_TO_TOKEN

Path to token.

PAYLOAD_LEN

Length of encryption payload.

RANDOMNESS_POOL 🔒

Randomness pool file path.

TAG_LEN

Length of encryption tag.

TOKEN_LENGTH_LIMIT 🔒

Token length limit to ensure we don’t bother decrypting huge character sequences. Tokens larger than this are automatically rejected. The value is computed based on the expected length of the base64 encoded Token struct including a small deviation.

_DERIVE_Display_FOR_Error 🔒