Module vmm::persist

Save/restore utilities. Defines state structures for saving/restoring a Firecracker microVM.

Structs

GuestRegionUffdMapping

This describes the mapping between Firecracker base virtual address and offset in the buffer or file backend for a guest memory region. It is used to tell an external process/thread where to populate the guest memory data for this range.

MicrovmState

Contains the necesary state for saving/restoring a microVM.

VmInfo

Holds information related to the VM that is not part of VmState.

Enums

CreateSnapshotError

Errors associated with creating a snapshot.

GuestMemoryFromFileError

Error type for guest_memory_from_file.

GuestMemoryFromUffdError

Error type for guest_memory_from_uffd

MicrovmStateError

Errors related to saving and restoring Microvm state.

RestoreFromSnapshotError

Error type for restore_from_snapshot.

RestoreFromSnapshotGuestMemoryError

Sub-Error type for restore_from_snapshot to contain either GuestMemoryFromFileError or GuestMemoryFromUffdError within RestoreFromSnapshotError.

SnapShotStateSanityCheckError

Error type for snapshot_state_sanity_check.

SnapshotStateFromFileError

Error type for snapshot_state_from_file

Constants

FC_V0_23_MAX_DEVICES 🔒

_DERIVE_Display_FOR_CreateSnapshotError 🔒

_DERIVE_Display_FOR_GuestMemoryFromFileError 🔒

_DERIVE_Display_FOR_GuestMemoryFromUffdError 🔒

_DERIVE_Display_FOR_MicrovmStateError 🔒

_DERIVE_Display_FOR_RestoreFromSnapshotError 🔒

_DERIVE_Display_FOR_RestoreFromSnapshotGuestMemoryError 🔒

_DERIVE_Display_FOR_SnapShotStateSanityCheckError 🔒

_DERIVE_Display_FOR_SnapshotStateFromFileError 🔒

Functions

create_snapshot

Creates a Microvm snapshot.

guest_memory_from_file 🔒

guest_memory_from_uffd 🔒

restore_from_snapshot

Loads a Microvm snapshot producing a ‘paused’ Microvm.

snapshot_memory_to_file 🔒

Takes a snapshot of the virtual machine running inside the given Vmm and saves it to mem_file_path.

snapshot_state_from_file 🔒

snapshot_state_sanity_check

Performs sanity checks against the state file and returns specific errors.

snapshot_state_to_file 🔒

validate_cpu_vendor

Validates that snapshot CPU vendor matches the host CPU vendor.