Struct vmm::rate_limiter::TokenBucket

pub struct TokenBucket {
    size: u64,
    initial_one_time_burst: u64,
    refill_time: u64,
    one_time_burst: u64,
    budget: u64,
    last_update: Instant,
    processed_capacity: u64,
    processed_refill_time: u64,
}

TokenBucket provides a lower level interface to rate limiting with a configurable capacity, refill-rate and initial burst.

Fields
size: u64initial_one_time_burst: u64refill_time: u64one_time_burst: u64budget: u64last_update: Instantprocessed_capacity: u64processed_refill_time: u64
Implementations

impl TokenBucket

pub fn new( size: u64, one_time_burst: u64, complete_refill_time_ms: u64 ) -> Option<Self>

Creates a TokenBucket wrapped in an Option.

TokenBucket created is of size total capacity and takes complete_refill_time_ms milliseconds to go from zero tokens to total capacity. The one_time_burst is initial extra credit on top of total capacity, that does not replenish and which can be used for an initial burst of data.

If the size or the complete refill time are zero, then None is returned.

fn auto_replenish(&mut self)

pub fn reduce(&mut self, tokens: u64) -> BucketReduction

Attempts to consume tokens from the bucket and returns whether the action succeeded.

pub fn force_replenish(&mut self, tokens: u64)

“Manually” adds tokens to bucket.

pub fn capacity(&self) -> u64

Returns the capacity of the token bucket.

pub fn one_time_burst(&self) -> u64

Returns the remaining one time burst budget.

pub fn refill_time_ms(&self) -> u64

Returns the time in milliseconds required to to completely fill the bucket.

pub fn budget(&self) -> u64

Returns the current budget (one time burst allowance notwithstanding).

pub fn initial_one_time_burst(&self) -> u64

Returns the initially configured one time burst budget.