Struct vmm::resources::VmResources

pub struct VmResources {
    pub vm_config: VmConfig,
    boot_source: BootSource,
    pub block: BlockBuilder,
    pub vsock: VsockBuilder,
    pub balloon: BalloonBuilder,
    pub net_builder: NetBuilder,
    pub entropy: EntropyDeviceBuilder,
    pub mmds: Option<Arc<Mutex<Mmds>>>,
    pub mmds_size_limit: usize,
    pub boot_timer: bool,
}

A data structure that encapsulates the device configurations held in the Vmm.

Fields
vm_config: VmConfig

The vCpu and memory configuration for this microVM.

boot_source: BootSource

The boot source spec (contains both config and builder) for this microVM.

block: BlockBuilder

The block devices.

vsock: VsockBuilder

The vsock device.

balloon: BalloonBuilder

The balloon device.

net_builder: NetBuilder

The network devices builder.

entropy: EntropyDeviceBuilder

The entropy device builder.

mmds: Option<Arc<Mutex<Mmds>>>

The optional Mmds data store.

mmds_size_limit: usize

Data store limit for the mmds.

boot_timer: bool

Whether or not to load boot timer device.

Implementations

impl VmResources

pub fn from_json( config_json: &str, instance_info: &InstanceInfo, mmds_size_limit: usize, metadata_json: Option<&str> ) -> Result<Self, ResourcesError>

Configures Vmm resources as described by the config_json param.

pub fn mmds_or_default(&mut self) -> &Arc<Mutex<Mmds>>

If not initialised, create the mmds data store with the default config.

pub fn locked_mmds_or_default(&mut self) -> MutexGuard<'_, Mmds>

If not initialised, create the mmds data store with the default config.

pub fn update_from_restored_device(&mut self, device: SharedDeviceType)

Updates the resources from a restored device (used for configuring resources when restoring from a snapshot).

pub fn track_dirty_pages(&self) -> bool

Returns whether dirty page tracking is enabled or not.

pub fn set_track_dirty_pages(&mut self, dirty_page_tracking: bool)

Configures the dirty page tracking functionality of the microVM.

pub fn set_custom_cpu_template(&mut self, cpu_template: CustomCpuTemplate)

Add a custom CPU template to the VM resources to configure vCPUs.

pub fn update_vm_config( &mut self, update: &MachineConfigUpdate ) -> Result<(), VmConfigError>

Updates the configuration of the microVM.

fn mmds_config(&self) -> Option<MmdsConfig>

pub fn boot_source_config(&self) -> &BootSourceConfig

Gets a reference to the boot source configuration.

pub fn boot_source_builder(&self) -> Option<&BootConfig>

Gets a reference to the boot source builder.

pub fn set_balloon_device( &mut self, config: BalloonDeviceConfig ) -> Result<(), BalloonConfigError>

Sets a balloon device to be attached when the VM starts.

pub fn build_boot_source( &mut self, boot_source_cfg: BootSourceConfig ) -> Result<(), BootSourceConfigError>

Obtains the boot source hooks (kernel fd, command line creation and validation).

pub fn set_boot_source_config(&mut self, boot_source_cfg: BootSourceConfig)

Set the boot source configuration (contains raw kernel config details).

pub fn set_block_device( &mut self, block_device_config: BlockDeviceConfig ) -> Result<(), DriveError>

Inserts a block to be attached when the VM starts.

pub fn build_net_device( &mut self, body: NetworkInterfaceConfig ) -> Result<(), NetworkInterfaceError>

Builds a network device to be attached when the VM starts.

pub fn set_vsock_device( &mut self, config: VsockDeviceConfig ) -> Result<(), VsockConfigError>

Sets a vsock device to be attached when the VM starts.

pub fn build_entropy_device( &mut self, body: EntropyDeviceConfig ) -> Result<(), EntropyDeviceError>

Builds an entropy device to be attached when the VM starts.

pub fn set_mmds_config( &mut self, config: MmdsConfig, instance_id: &str ) -> Result<(), MmdsConfigError>

Setter for mmds config.

pub fn set_mmds_version( &mut self, version: MmdsVersion, instance_id: &str ) -> Result<(), MmdsConfigError>

Updates MMDS version.

fn set_mmds_network_stack_config( &mut self, config: &MmdsConfig ) -> Result<(), MmdsConfigError>