Struct vmm::rpc_interface::PrebootApiController

pub struct PrebootApiController<'a> {
    seccomp_filters: &'a BpfThreadMap,
    instance_info: InstanceInfo,
    vm_resources: &'a mut VmResources,
    event_manager: &'a mut EventManager,
    built_vmm: Option<Arc<Mutex<Vmm>>>,
    boot_path: bool,
    fatal_error: Option<BuildMicrovmFromRequestsError>,
}

Enables pre-boot setup and instantiation of a Firecracker VMM.

Fields
seccomp_filters: &'a BpfThreadMapinstance_info: InstanceInfovm_resources: &'a mut VmResourcesevent_manager: &'a mut EventManagerbuilt_vmm: Option<Arc<Mutex<Vmm>>>boot_path: boolfatal_error: Option<BuildMicrovmFromRequestsError>
Implementations

impl<'a> PrebootApiController<'a>

pub fn new( seccomp_filters: &'a BpfThreadMap, instance_info: InstanceInfo, vm_resources: &'a mut VmResources, event_manager: &'a mut EventManager ) -> Self

Constructor for the PrebootApiController.

pub fn build_microvm_from_requests( seccomp_filters: &BpfThreadMap, event_manager: &mut EventManager, instance_info: InstanceInfo, from_api: &Receiver<ApiRequest>, to_api: &Sender<ApiResponse>, api_event_fd: &EventFd, boot_timer_enabled: bool, mmds_size_limit: usize, metadata_json: Option<&str> ) -> Result<(VmResources, Arc<Mutex<Vmm>>), BuildMicrovmFromRequestsError>

Default implementation for the function that builds and starts a microVM.

Returns a populated VmResources object and a running Vmm object.

pub fn handle_preboot_request( &mut self, request: VmmAction ) -> Result<VmmData, VmmActionError>

Handles the incoming preboot request and provides a response for it. Returns a built/running Vmm after handling a successful StartMicroVm request.

fn balloon_config(&mut self) -> Result<VmmData, VmmActionError>

fn insert_block_device( &mut self, cfg: BlockDeviceConfig ) -> Result<VmmData, VmmActionError>

fn insert_net_device( &mut self, cfg: NetworkInterfaceConfig ) -> Result<VmmData, VmmActionError>

fn set_balloon_device( &mut self, cfg: BalloonDeviceConfig ) -> Result<VmmData, VmmActionError>

fn set_boot_source( &mut self, cfg: BootSourceConfig ) -> Result<VmmData, VmmActionError>

fn set_mmds_config( &mut self, cfg: MmdsConfig ) -> Result<VmmData, VmmActionError>

fn update_vm_config( &mut self, cfg: MachineConfigUpdate ) -> Result<VmmData, VmmActionError>

fn set_custom_cpu_template( &mut self, cpu_template: CustomCpuTemplate ) -> Result<VmmData, VmmActionError>

fn set_vsock_device( &mut self, cfg: VsockDeviceConfig ) -> Result<VmmData, VmmActionError>

fn set_entropy_device( &mut self, cfg: EntropyDeviceConfig ) -> Result<VmmData, VmmActionError>

fn start_microvm(&mut self) -> Result<VmmData, VmmActionError>

fn load_snapshot( &mut self, load_params: &LoadSnapshotParams ) -> Result<VmmData, LoadSnapshotError>