Trait vmm::rpc_interface::MmdsRequestHandler

trait MmdsRequestHandler {
    // Required method
    fn mmds(&mut self) -> MutexGuard<'_, Mmds>;

    // Provided methods
    fn get_mmds(&mut self) -> Result<VmmData, VmmActionError> { ... }
    fn patch_mmds(&mut self, value: Value) -> Result<VmmData, VmmActionError> { ... }
    fn put_mmds(&mut self, value: Value) -> Result<VmmData, VmmActionError> { ... }
}

Trait used for deduplicating the MMDS request handling across the two ApiControllers. The methods get a mutable reference to self because the methods should initialise the data store with the defaults if it’s not already initialised.

Required Methods

fn mmds(&mut self) -> MutexGuard<'_, Mmds>

Provided Methods

fn get_mmds(&mut self) -> Result<VmmData, VmmActionError>

fn patch_mmds(&mut self, value: Value) -> Result<VmmData, VmmActionError>

fn put_mmds(&mut self, value: Value) -> Result<VmmData, VmmActionError>