Module vmm::vmm_config::balloon

Wrapper for configuring the balloon device.

Re-exports

pub use crate::devices::virtio::balloon::device::BalloonStats;

pub use crate::devices::virtio::balloon::BALLOON_DEV_ID;

Structs

BalloonBuilder

A builder for Balloon devices from ‘BalloonDeviceConfig’.

BalloonDeviceConfig

This struct represents the strongly typed equivalent of the json body from balloon related requests.

BalloonUpdateConfig

The data fed into a balloon update request. Currently, only the number of pages and the stats polling interval can be updated.

BalloonUpdateStatsConfig

The data fed into a balloon statistics interval update request. Note that the state of the statistics cannot be changed from ON to OFF or vice versa after boot, only the interval of polling can be changed if the statistics were activated in the device configuration.

Enums

BalloonConfigError

Errors associated with the operations allowed on the balloon.

Constants

_DERIVE_Display_FOR_BalloonConfigError 🔒

Type Aliases

MutexBalloon 🔒