Module vmm::vmm_config::machine_config

Wrapper for configuring the memory and CPU of the microVM.

Structs

MachineConfig

Struct used in PUT /machine-config API call.

MachineConfigUpdate

Struct used in PATCH /machine-config API call. Used to update VmConfig in VmResources. This struct mirrors all the fields in MachineConfig. All fields are optional, but at least one needs to be specified. If a field is Some(value) then we assume an update is requested for that field.

VmConfig

Configuration of the microvm.

Enums

VmConfigError

Errors associated with configuring the microVM.

Constants

DEFAULT_MEM_SIZE_MIB

The default memory size of the VM, in MiB.

MAX_SUPPORTED_VCPUS

Firecracker aims to support small scale workloads only, so limit the maximum vCPUs supported.

_DERIVE_Display_FOR_VmConfigError 🔒

Functions

deserialize_smt 🔒

Deserialization function for the smt field in MachineConfig and MachineConfigUpdate. This is called only when smt is present in the JSON configuration.

deserialize_vcpu_num 🔒

Deserialization function for the vcpu_num field in MachineConfig and MachineConfigUpdate. This is called only when vcpu_num is present in the JSON configuration. T can be either u8 or Option<u8> which both support ordering if vcpu_num is present in the JSON.