Module vmm::vmm_config::snapshot

Wrapper for configuring microVM snapshots and the microVM state. Configurations used in the snapshotting context.

Structs

CreateSnapshotParams

Stores the configuration that will be used for creating a snapshot.

LoadSnapshotConfig

Stores the configuration for loading a snapshot that is provided by the user.

LoadSnapshotParams

Stores the configuration that will be used for loading a snapshot.

MemBackendConfig

Stores the configuration used for managing snapshot memory.

Version

For crates that depend on vmm we export. SemVer version as defined by https://semver.org.

Vm

Keeps the microVM state necessary in the snapshotting context.

Enums

MemBackendType

Specifies the method through which guest memory will get populated when resuming from a snapshot:

SnapshotType

The snapshot type options that are available when creating a new snapshot.

VmState

The microVM state options.