Struct vmm::vstate::memory::MmapRegion

pub struct MmapRegion<B = ()> {
    addr: *mut u8,
    size: usize,
    bitmap: B,
    file_offset: Option<FileOffset>,
    prot: i32,
    flags: i32,
    owned: bool,
    hugetlbfs: Option<bool>,
}

Helper structure for working with mmaped memory regions in Unix.

The structure is used for accessing the guest’s physical memory by mmapping it into the current process.

Limitations

When running a 64-bit virtual machine on a 32-bit hypervisor, only part of the guest’s physical memory may be mapped into the current process due to the limited virtual address space size of the process.

Fields
addr: *mut u8size: usizebitmap: Bfile_offset: Option<FileOffset>prot: i32flags: i32owned: boolhugetlbfs: Option<bool>
Implementations

impl<B> MmapRegion<B>
where B: NewBitmap,

pub fn new(size: usize) -> Result<MmapRegion<B>, Error>

Creates a shared anonymous mapping of size bytes.

Arguments

size - The size of the memory region in bytes.

pub fn from_file( file_offset: FileOffset, size: usize ) -> Result<MmapRegion<B>, Error>

Creates a shared file mapping of size bytes.

Arguments

file_offset - The mapping will be created at offset file_offset.start in the file referred to by file_offset.file.

size - The size of the memory region in bytes.

pub fn build( file_offset: Option<FileOffset>, size: usize, prot: i32, flags: i32 ) -> Result<MmapRegion<B>, Error>

Creates a mapping based on the provided arguments.

Arguments

file_offset - if provided, the method will create a file mapping at offset file_offset.start in the file referred to by file_offset.file.

size - The size of the memory region in bytes.

prot - The desired memory protection of the mapping.

flags - This argument determines whether updates to the mapping are visible to other processes mapping the same region, and whether updates are carried through to the underlying file.

pub unsafe fn build_raw( addr: *mut u8, size: usize, prot: i32, flags: i32 ) -> Result<MmapRegion<B>, Error>

Creates a MmapRegion instance for an externally managed mapping.

This method is intended to be used exclusively in situations in which the mapping backing the region is provided by an entity outside the control of the caller (e.g. the dynamic linker).

Arguments

addr - Pointer to the start of the mapping. Must be page-aligned.

size - The size of the memory region in bytes.

prot - Must correspond to the memory protection attributes of the existing mapping.

flags - Must correspond to the flags that were passed to mmap for the creation of the existing mapping.

Safety

To use this safely, the caller must guarantee that addr and size define a region within a valid mapping that is already present in the process.

impl<B> MmapRegion<B>
where B: Bitmap,

pub fn as_ptr(&self) -> *mut u8

Returns a pointer to the beginning of the memory region. Mutable accesses performed using the resulting pointer are not automatically accounted for by the dirty bitmap tracking functionality.

Should only be used for passing this region to ioctls for setting guest memory.

pub fn size(&self) -> usize

Returns the size of this region.

pub fn file_offset(&self) -> Option<&FileOffset>

Returns information regarding the offset into the file backing this region (if any).

pub fn prot(&self) -> i32

Returns the value of the prot parameter passed to mmap when mapping this region.

pub fn flags(&self) -> i32

Returns the value of the flags parameter passed to mmap when mapping this region.

pub fn owned(&self) -> bool

Returns true if the mapping is owned by this MmapRegion instance.

pub fn fds_overlap<T>(&self, other: &MmapRegion<T>) -> bool
where T: Bitmap,

Checks whether this region and other are backed by overlapping FileOffset objects.

This is mostly a sanity check available for convenience, as different file descriptors can alias the same file.

pub fn set_hugetlbfs(&mut self, hugetlbfs: bool)

Set the hugetlbfs of the region

pub fn is_hugetlbfs(&self) -> Option<bool>

Returns true if the region is hugetlbfs

pub fn bitmap(&self) -> &B

Returns a reference to the inner bitmap object.