Trait vmm::vstate::memory::Bytes

pub trait Bytes<A> {
    type E;

    // Required methods
    fn write(&self, buf: &[u8], addr: A) -> Result<usize, Self::E>;
    fn read(&self, buf: &mut [u8], addr: A) -> Result<usize, Self::E>;
    fn write_slice(&self, buf: &[u8], addr: A) -> Result<(), Self::E>;
    fn read_slice(&self, buf: &mut [u8], addr: A) -> Result<(), Self::E>;
    fn read_from<F>(
        &self,
        addr: A,
        src: &mut F,
        count: usize
    ) -> Result<usize, Self::E>
       where F: Read;
    fn read_exact_from<F>(
        &self,
        addr: A,
        src: &mut F,
        count: usize
    ) -> Result<(), Self::E>
       where F: Read;
    fn write_to<F>(
        &self,
        addr: A,
        dst: &mut F,
        count: usize
    ) -> Result<usize, Self::E>
       where F: Write;
    fn write_all_to<F>(
        &self,
        addr: A,
        dst: &mut F,
        count: usize
    ) -> Result<(), Self::E>
       where F: Write;
    fn store<T>(&self, val: T, addr: A, order: Ordering) -> Result<(), Self::E>
       where T: AtomicAccess;
    fn load<T>(&self, addr: A, order: Ordering) -> Result<T, Self::E>
       where T: AtomicAccess;

    // Provided methods
    fn write_obj<T>(&self, val: T, addr: A) -> Result<(), Self::E>
       where T: ByteValued { ... }
    fn read_obj<T>(&self, addr: A) -> Result<T, Self::E>
       where T: ByteValued { ... }
}

A container to host a range of bytes and access its content.

Candidates which may implement this trait include:

anonymous memory areas

mmapped memory areas

data files

a proxy to access memory on remote

Required Associated Types

type E

Associated error codes

Required Methods

fn write(&self, buf: &[u8], addr: A) -> Result<usize, Self::E>

Writes a slice into the container at addr.

Returns the number of bytes written. The number of bytes written can be less than the length of the slice if there isn’t enough room in the container.

fn read(&self, buf: &mut [u8], addr: A) -> Result<usize, Self::E>

Reads data from the container at addr into a slice.

Returns the number of bytes read. The number of bytes read can be less than the length of the slice if there isn’t enough data within the container.

fn write_slice(&self, buf: &[u8], addr: A) -> Result<(), Self::E>

Writes the entire content of a slice into the container at addr.

Errors

Returns an error if there isn’t enough space within the container to write the entire slice. Part of the data may have been copied nevertheless.

fn read_slice(&self, buf: &mut [u8], addr: A) -> Result<(), Self::E>

Reads data from the container at addr to fill an entire slice.

Errors

Returns an error if there isn’t enough data within the container to fill the entire slice. Part of the data may have been copied nevertheless.

fn read_from<F>( &self, addr: A, src: &mut F, count: usize ) -> Result<usize, Self::E>
where F: Read,

👎Deprecated: Use .read_volatile_from or the functions of the ReadVolatile trait instead

Reads up to count bytes from an object and writes them into the container at addr.

Returns the number of bytes written into the container.

Arguments

addr - Begin writing at this address.

src - Copy from src into the container.

count - Copy count bytes from src into the container.

fn read_exact_from<F>( &self, addr: A, src: &mut F, count: usize ) -> Result<(), Self::E>
where F: Read,

👎Deprecated: Use .read_exact_volatile_from or the functions of the ReadVolatile trait instead

Reads exactly count bytes from an object and writes them into the container at addr.

Errors

Returns an error if count bytes couldn’t have been copied from src to the container. Part of the data may have been copied nevertheless.

Arguments

addr - Begin writing at this address.

src - Copy from src into the container.

count - Copy exactly count bytes from src into the container.

fn write_to<F>( &self, addr: A, dst: &mut F, count: usize ) -> Result<usize, Self::E>
where F: Write,

👎Deprecated: Use .write_volatile_to or the functions of the WriteVolatile trait instead

Reads up to count bytes from the container at addr and writes them it into an object.

Returns the number of bytes written into the object.

Arguments

addr - Begin reading from this address.

dst - Copy from the container to dst.

count - Copy count bytes from the container to dst.

fn write_all_to<F>( &self, addr: A, dst: &mut F, count: usize ) -> Result<(), Self::E>
where F: Write,

👎Deprecated: Use .write_all_volatile_to or the functions of the WriteVolatile trait instead

Reads exactly count bytes from the container at addr and writes them into an object.

Errors

Returns an error if count bytes couldn’t have been copied from the container to dst. Part of the data may have been copied nevertheless.

Arguments

addr - Begin reading from this address.

dst - Copy from the container to dst.

count - Copy exactly count bytes from the container to dst.

fn store<T>(&self, val: T, addr: A, order: Ordering) -> Result<(), Self::E>
where T: AtomicAccess,

Atomically store a value at the specified address.

fn load<T>(&self, addr: A, order: Ordering) -> Result<T, Self::E>
where T: AtomicAccess,

Atomically load a value from the specified address.

Provided Methods

fn write_obj<T>(&self, val: T, addr: A) -> Result<(), Self::E>
where T: ByteValued,

Writes an object into the container at addr.

Errors

Returns an error if the object doesn’t fit inside the container.

fn read_obj<T>(&self, addr: A) -> Result<T, Self::E>
where T: ByteValued,

Reads an object from the container at addr.

Reading from a volatile area isn’t strictly safe as it could change mid-read. However, as long as the type T is plain old data and can handle random initialization, everything will be OK.

Errors

Returns an error if there’s not enough data inside the container.

Object Safety

This trait is not object safe.