Struct vmm::vstate::vcpu::aarch64::KvmVcpu

pub struct KvmVcpu {
    pub index: u8,
    pub fd: VcpuFd,
    pub mmio_bus: Option<Bus>,
    /* private fields */
}

A wrapper around creating and using a kvm aarch64 vcpu.

Fields
index: u8

Index of vcpu.

fd: VcpuFd

KVM vcpu fd.

mmio_bus: Option<Bus>

Mmio bus.

Implementations

impl KvmVcpu

pub fn new(index: u8, vm: &Vm) -> Result<Self, KvmVcpuError>

Constructs a new kvm vcpu with arch specific functionality.

Arguments

index - Represents the 0-based CPU index between [0, max vcpus).

vm - The vm to which this vcpu will get attached.

pub fn get_mpidr(&self) -> u64

Gets the MPIDR register value.

pub fn configure( &mut self, guest_mem: &GuestMemoryMmap, kernel_load_addr: GuestAddress, vcpu_config: &VcpuConfig ) -> Result<(), KvmVcpuError>

Configures an aarch64 specific vcpu for booting Linux.

Arguments

guest_mem - The guest memory used by this microvm.

kernel_load_addr - Offset from guest_mem at which the kernel is loaded.

vcpu_config - The vCPU configuration.

pub fn init( &mut self, vm_fd: &VmFd, vcpu_features: &[VcpuFeatures] ) -> Result<(), KvmVcpuError>

Initializes an aarch64 specific vcpu for booting Linux.

Arguments

vm_fd - The kvm VmFd for this microvm.

pub fn default_kvi( vm_fd: &VmFd, index: u8 ) -> Result<kvm_vcpu_init, KvmVcpuError>

Creates default kvi struct based on vcpu index.

pub fn save_state(&self) -> Result<VcpuState, KvmVcpuError>

Save the KVM internal state.

pub fn restore_state( &mut self, vm_fd: &VmFd, state: &VcpuState ) -> Result<(), KvmVcpuError>

Use provided state to populate KVM internal state.

pub fn dump_cpu_config(&self) -> Result<CpuConfiguration, KvmVcpuError>

Dumps CPU configuration.

pub fn run_arch_emulation( &self, exit: VcpuExit<'_> ) -> Result<VcpuEmulation, VcpuError>

Runs the vCPU in KVM context and handles the kvm exit reason.

Returns error or enum specifying whether emulation was handled or interrupted.