Struct vmm::vstate::vcpu::Vcpu

pub struct Vcpu {
    pub kvm_vcpu: KvmVcpu,
    exit_evt: EventFd,
    event_receiver: Receiver<VcpuEvent>,
    event_sender: Option<Sender<VcpuEvent>>,
    response_receiver: Option<Receiver<VcpuResponse>>,
    response_sender: Sender<VcpuResponse>,
}

A wrapper around creating and using a vcpu.

Fields
kvm_vcpu: KvmVcpu

Access to kvm-arch specific functionality.

exit_evt: EventFd

File descriptor for vcpu to trigger exit event on vmm.

event_receiver: Receiver<VcpuEvent>

The receiving end of events channel owned by the vcpu side.

event_sender: Option<Sender<VcpuEvent>>

The transmitting end of the events channel which will be given to the handler.

response_receiver: Option<Receiver<VcpuResponse>>

The receiving end of the responses channel which will be given to the handler.

response_sender: Sender<VcpuResponse>

The transmitting end of the responses channel owned by the vcpu side.

Implementations

impl Vcpu

const TLS_VCPU_PTR: LocalKey<Cell<Option<*const Vcpu>>> = _

fn init_thread_local_data(&mut self) -> Result<(), VcpuError>

Associates self with the current thread.

It is a prerequisite to successfully run init_thread_local_data() before using run_on_thread_local() on the current thread. This function will return an error if there already is a Vcpu present in the TLS.

fn reset_thread_local_data(&mut self) -> Result<(), VcpuError>

Deassociates self from the current thread.

Should be called if the current self had called init_thread_local_data() and now needs to move to a different thread.

Fails if self was not previously associated with the current thread.

unsafe fn run_on_thread_local<F>(func: F) -> Result<(), VcpuError>
where F: FnOnce(&Vcpu),

Runs func for the Vcpu associated with the current thread.

It requires that init_thread_local_data() was run on this thread.

Fails if there is no Vcpu associated with the current thread.

Safety

This is marked unsafe as it allows temporary aliasing through dereferencing from pointer an already borrowed Vcpu.

pub fn register_kick_signal_handler()

Registers a signal handler which makes use of TLS and kvm immediate exit to kick the vcpu running on the current thread, if there is one.

pub fn new(index: u8, vm: &Vm, exit_evt: EventFd) -> Result<Self, VcpuError>

Constructs a new VCPU for vm.

Arguments

index - Represents the 0-based CPU index between [0, max vcpus).

vm - The vm to which this vcpu will get attached.

exit_evt - An EventFd that will be written into when this vcpu exits.

pub fn set_mmio_bus(&mut self, mmio_bus: Bus)

Sets a MMIO bus for this vcpu.

pub fn start_threaded( self, seccomp_filter: Arc<BpfProgram>, barrier: Arc<Barrier> ) -> Result<VcpuHandle, StartThreadedError>

Moves the vcpu to its own thread and constructs a VcpuHandle. The handle can be used to control the remote vcpu.

pub fn run(&mut self, seccomp_filter: BpfProgramRef<'_>)

Main loop of the vCPU thread.

Runs the vCPU in KVM context in a loop. Handles KVM_EXITs then goes back in. Note that the state of the VCPU and associated VM must be setup first for this to do anything useful.

fn running(&mut self) -> StateMachine<Self>

fn paused(&mut self) -> StateMachine<Self>

fn exit(&mut self, exit_code: FcExitCode) -> StateMachine<Self>

pub fn emulate(&self) -> Result<VcpuExit<'_>, Error>

Calls KVM_RUN with this Vcpu’s underlying file descriptor.

Blocks until a VM_EXIT is received, in which case this function returns a [VcpuExit] containing the reason.

pub fn run_emulation(&self) -> Result<VcpuEmulation, VcpuError>

Runs the vCPU in KVM context and handles the kvm exit reason.

Returns error or enum specifying whether emulation was handled or interrupted.