Struct vmm::vstate::vcpu::x86_64::KvmVcpu

pub struct KvmVcpu {
    pub index: u8,
    pub fd: VcpuFd,
    pub pio_bus: Option<Bus>,
    pub mmio_bus: Option<Bus>,
    msrs_to_save: HashSet<u32>,
}

A wrapper around creating and using a kvm x86_64 vcpu.

Fields
index: u8

Index of vcpu.

fd: VcpuFd

KVM vcpu fd.

pio_bus: Option<Bus>

Pio bus.

mmio_bus: Option<Bus>

Mmio bus.

msrs_to_save: HashSet<u32>
Implementations

impl KvmVcpu

pub fn new(index: u8, vm: &Vm) -> Result<Self, KvmVcpuError>

Constructs a new kvm vcpu with arch specific functionality.

Arguments

index - Represents the 0-based CPU index between [0, max vcpus).

vm - The vm to which this vcpu will get attached.

pub fn configure( &mut self, guest_mem: &GuestMemoryMmap, kernel_start_addr: GuestAddress, vcpu_config: &VcpuConfig ) -> Result<(), KvmVcpuConfigureError>

Configures a x86_64 specific vcpu for booting Linux and should be called once per vcpu.

Arguments

guest_mem - The guest memory used by this microvm.

kernel_start_addr - Offset from guest_mem at which the kernel starts.

vcpu_config - The vCPU configuration.

cpuid - The capabilities exposed by this vCPU.

pub fn set_pio_bus(&mut self, pio_bus: Bus)

Sets a Port Mapped IO bus for this vcpu.

pub fn get_tsc_khz(&self) -> Result<u32, GetTscError>

Get the current TSC frequency for this vCPU.

Errors

When [kvm_ioctls::VcpuFd::get_tsc_khz] errrors.

fn get_cpuid(&self) -> Result<CpuId, KvmVcpuError>

Get CPUID for this vCPU.

Opposed to KVM_GET_SUPPORTED_CPUID, KVM_GET_CPUID2 does not update “nent” with valid number of entries on success. Thus, when it passes “num_entries” greater than required, zeroed entries follow after valid entries. This function removes such zeroed empty entries.

Errors

When [kvm_ioctls::VcpuFd::get_cpuid2] returns errors.

fn get_msr_chunks( &self, msr_index_list: &[u32] ) -> Result<Vec<Msrs>, KvmVcpuError>

Get MSR chunks for the given MSR index list.

KVM only supports getting KVM_MAX_MSR_ENTRIES at a time, so we divide the list of MSR indices into chunks, call KVM_GET_MSRS for each chunk, and collect into a Vec<Msrs>.

Arguments

msr_index_list: List of MSR indices.

Errors

When [kvm_bindings::Msrs::new] returns errors.

When [kvm_ioctls::VcpuFd::get_msrs] returns errors.

When the return value of [kvm_ioctls::VcpuFd::get_msrs] (the number of entries that could be gotten) is less than expected.

pub fn get_msrs( &self, msr_index_list: &[u32] ) -> Result<HashMap<u32, u64>, KvmVcpuError>

Get MSRs for the given MSR index list.

Arguments

msr_index_list: List of MSR indices

Errors

When KvmVcpu::get_msr_chunks() returns errors.

pub fn save_state(&self) -> Result<VcpuState, KvmVcpuError>

Save the KVM internal state.

pub fn dump_cpu_config(&self) -> Result<CpuConfiguration, KvmVcpuError>

Dumps CPU configuration (CPUID and MSRs).

Opposed to save_state(), this dumps all the supported and dumpable MSRs not limited to serializable ones.

pub fn is_tsc_scaling_required( &self, state_tsc_freq: u32 ) -> Result<bool, GetTscError>

Checks whether the TSC needs scaling when restoring a snapshot.

Errors

When

pub fn set_tsc_khz(&self, tsc_freq: u32) -> Result<(), SetTscError>

Scale the TSC frequency of this vCPU to the one provided as a parameter.

pub fn restore_state(&self, state: &VcpuState) -> Result<(), KvmVcpuError>

Use provided state to populate KVM internal state.

pub fn run_arch_emulation( &self, exit: VcpuExit<'_> ) -> Result<VcpuEmulation, VcpuError>

Runs the vCPU in KVM context and handles the kvm exit reason.

Returns error or enum specifying whether emulation was handled or interrupted.