Struct vmm::vstate::vm::Vm

pub struct Vm {
    pub kvm_cap_modifiers: Vec<KvmCapability>,
    /* private fields */
}

A wrapper around creating and using a VM.

Fields
kvm_cap_modifiers: Vec<KvmCapability>

Additional capabilities that were specified in cpu template.

Implementations

impl Vm

Contains Vm functions that are usable across CPU architectures

pub fn new(kvm_cap_modifiers: Vec<KvmCapability>) -> Result<Self, VmError>

Constructs a new Vm using the given Kvm instance.

pub fn memory_init( &self, guest_mem: &GuestMemoryMmap, track_dirty_pages: bool ) -> Result<(), VmError>

Initializes the guest memory.

pub fn fd(&self) -> &VmFd

Gets a reference to the kvm file descriptor owned by this VM.

impl Vm

pub fn setup_irqchip(&mut self, vcpu_count: u8) -> Result<(), VmError>

Creates the GIC (Global Interrupt Controller).

pub fn get_irqchip(&self) -> &GICDevice

Gets a reference to the irqchip of the VM.

pub fn save_state(&self, mpidrs: &[u64]) -> Result<VmState, VmError>

Saves and returns the Kvm Vm state.

pub fn restore_state( &mut self, mpidrs: &[u64], state: &VmState ) -> Result<(), RestoreStateError>

Restore the KVM VM state

Errors

When GICDevice::restore_device errors.