Module vmm::arch

Implements platform specific functionality. Supported platforms: x86_64 and aarch64.

Re-exports

pub use crate::arch::x86_64::arch_memory_regions;

pub use crate::arch::x86_64::configure_system;

pub use crate::arch::x86_64::get_kernel_start;

pub use crate::arch::x86_64::initrd_load_addr;

pub use crate::arch::x86_64::layout::CMDLINE_MAX_SIZE;

pub use crate::arch::x86_64::layout::IRQ_BASE;

pub use crate::arch::x86_64::layout::IRQ_MAX;

pub use crate::arch::x86_64::ConfigurationError;

pub use crate::arch::x86_64::MMIO_MEM_SIZE;

pub use crate::arch::x86_64::MMIO_MEM_START;

Modules

x86_64

Module for x86_64 related functionality.

Structs

InitrdConfig

Type for passing information about the initrd in the guest memory.

Enums

DeviceType

Types of devices that can get attached to this platform.

Constants

PAGE_SIZE

Default (smallest) memory page size for the supported architectures.