Module vmm::devices::virtio::balloon

Implements a virtio balloon device.

Re-exports

pub use self::device::Balloon;

pub use self::device::BalloonConfig;

pub use self::device::BalloonStats;

Modules

device

metrics

Defines the metrics system for balloon devices.

persist

Defines the structures needed for saving/restoring balloon devices.

Enums

BalloonError

Balloon device related errors.

RemoveRegionError

Constants

BALLOON_CONFIG_SPACE_SIZE

The size of the config space.

BALLOON_DEV_ID

Device ID used in MMIO device identification. Because Balloon is unique per-vm, this ID can be hardcoded.

BALLOON_NUM_QUEUES

Number of virtio queues.

BALLOON_QUEUE_SIZES

Virtio queue sizes, in number of descriptor chain heads.

DEFLATE_INDEX

The index of the deflate queue from Balloon device queues/queues_evts vector.

INFLATE_INDEX

The index of the deflate queue from Balloon device queues/queues_evts vector.

MAX_PAGES_IN_DESC

The maximum number of pages that can be received in a single descriptor.

MAX_PAGE_COMPACT_BUFFER

The maximum number of pages that can be compacted into ranges during process_inflate(). Needs to be a multiple of MAX_PAGES_IN_DESC.

MIB_TO_4K_PAGES

STATS_INDEX

The index of the deflate queue from Balloon device queues/queues_evts vector.

VIRTIO_BALLOON_PFN_SHIFT

The addresses given by the driver are divided by 4096.