Struct vmm::devices::virtio::virtio_block::device::VirtioBlock

pub struct VirtioBlock {

    pub avail_features: u64,
    pub acked_features: u64,
    pub config_space: Vec<u8>,
    pub activate_evt: EventFd,
    pub queues: Vec<Queue>,
    pub queue_evts: [EventFd; 1],
    pub device_state: DeviceState,
    pub irq_trigger: IrqTrigger,
    pub id: String,
    pub partuuid: Option<String>,
    pub cache_type: CacheType,
    pub root_device: bool,
    pub read_only: bool,
    pub disk: DiskProperties,
    pub rate_limiter: RateLimiter,
    pub is_io_engine_throttled: bool,
    pub metrics: Arc<BlockDeviceMetrics>,

}

Virtio device for exposing block level read/write operations on a host file.

Fields
avail_features: u64acked_features: u64config_space: Vec<u8>activate_evt: EventFdqueues: Vec<Queue>queue_evts: [EventFd; 1]device_state: DeviceStateirq_trigger: IrqTriggerid: Stringpartuuid: Option<String>cache_type: CacheTyperoot_device: boolread_only: booldisk: DiskPropertiesrate_limiter: RateLimiteris_io_engine_throttled: boolmetrics: Arc<BlockDeviceMetrics>
Implementations

impl VirtioBlock

pub fn new(config: VirtioBlockConfig) -> Result<VirtioBlock, VirtioBlockError>

Create a new virtio block device that operates on the given file.

The given file must be seekable and sizable.

pub fn config(&self) -> VirtioBlockConfig

Returns a copy of a device config

pub fn process_virtio_queues(&mut self)

Process device virtio queue(s).

pub fn process_queue(&mut self, queue_index: usize)

Device specific function for peaking inside a queue and processing descriptors.

pub fn process_async_completion_event(&mut self)

pub fn update_disk_image( &mut self, disk_image_path: String ) -> Result<(), VirtioBlockError>

Update the backing file and the config space of the block device.

pub fn update_rate_limiter(&mut self, bytes: BucketUpdate, ops: BucketUpdate)

Updates the parameters for the rate limiter

pub fn file_engine_type(&self) -> FileEngineType

Retrieve the file engine type.

pub fn prepare_save(&mut self)

Prepare device for being snapshotted.