Module vmm::dumbo::tcp

Provides functionality for handling incoming TCP connections.

Modules

connection

This module contains a minimalist TCP Connection implementation, which only supports passive open scenarios, and some auxiliary logic and data structures.

handler

Exposes simple TCP over IPv4 listener functionality via the TcpIPv4Handler structure.

Enums

NextSegmentStatus

Describes whether a particular entity (a Connection for example) has segments to send.

RstConfig

Represents the configuration of the sequence number and ACK number fields for outgoing RST segments.

Constants

MAX_WINDOW_SIZE

The largest possible window size (requires the window scaling option).

MSS_DEFAULT

The default maximum segment size (MSS) value, used when no MSS information is carried over the initial handshake.

Functions

seq_after

Returns true if a comes after b in the sequence number space, relative to the maximum possible window size.

seq_at_or_after

Returns true if a comes after, or is at b in the sequence number space, relative to the maximum possible window size.