Struct vmm::dumbo::tcp::handler::TcpIPv4Handler

pub struct TcpIPv4Handler {
    local_ipv4_addr: Ipv4Addr,
    local_port: u16,
    connections: HashMap<ConnectionTuple, Endpoint>,
    max_connections: usize,
    active_connections: HashSet<ConnectionTuple>,
    next_timeout: Option<(u64, ConnectionTuple)>,
    rst_queue: Vec<(ConnectionTuple, RstConfig)>,
    max_pending_resets: usize,
}

Implements a minimalist TCP over IPv4 listener.

Forwards incoming TCP segments to the appropriate connection object, based on the associated tuple, or attempts to establish new connections (when receiving SYN segments). Aside from constructors, the handler operation is based on three methods:

receive_packet examines an incoming IPv4 packet. It checks whether the destination address is correct, the attempts examine the inner TCP segment, making sure the destination port number is also correct. Then, it steers valid segments towards exiting connections, creates new connections for incoming SYN segments, and enqueues RST replies in response to any segments which cannot be associated with a connection (except other RST segments). On success, also describes any internal status changes triggered by the reception of the packet.

write_next_packet writes the next IPv4 packet (if available) that would be sent by the handler itself (right now it can only mean an enqueued RST), or one of the existing connections. On success, also describes any internal status changes triggered as the packet gets transmitted.

next_segment_status describes whether the handler can send a packet immediately, or after some retransmission timeout associated with a connection fires, or if there’s nothing to send for the moment. This is used to determine whether it’s appropriate to call write_next_packet.

Fields
local_ipv4_addr: Ipv4Addrlocal_port: u16connections: HashMap<ConnectionTuple, Endpoint>max_connections: usizeactive_connections: HashSet<ConnectionTuple>next_timeout: Option<(u64, ConnectionTuple)>rst_queue: Vec<(ConnectionTuple, RstConfig)>max_pending_resets: usize
Implementations

impl TcpIPv4Handler

pub fn new( local_ipv4_addr: Ipv4Addr, local_port: u16, max_connections: NonZeroUsize, max_pending_resets: NonZeroUsize ) -> Self

Creates a new TcpIPv4Handler.

The handler acts as if bound to local_addr:local_port, and will accept at most max_connections concurrent connections. RST segments generated by unexpected incoming segments are placed in a queue which is at most max_pending_resets long.

pub fn set_local_ipv4_addr(&mut self, ipv4_addr: Ipv4Addr)

Setter for the local IPv4 address of this TCP handler.

pub fn local_ipv4_addr(&self) -> Ipv4Addr

Returns the local IPv4 address of this TCP handler.

pub fn local_port(&self) -> u16

Returns the local port of this TCP handler.

pub fn max_connections(&self) -> usize

Returns the max connections of this TCP handler.

pub fn max_pending_resets(&self) -> usize

Returns the max pending resets of this TCP handler.

pub fn receive_packet<T: NetworkBytes + Debug, F: FnOnce(Request) -> Response>( &mut self, packet: &IPv4Packet<'_, T>, callback: F ) -> Result<RecvEvent, RecvError>

Contains logic for handling incoming segments.

Any changes to the state of the handler are communicated through an Ok(RecvEvent).

fn check_timeout(&mut self, value: u64, tuple: ConnectionTuple)

fn find_next_timeout(&mut self)

fn check_next_segment_status( &mut self, tuple: ConnectionTuple, status: NextSegmentStatus ) -> bool

fn add_connection(&mut self, tuple: ConnectionTuple, endpoint: Endpoint)

fn remove_connection(&mut self, tuple: ConnectionTuple)

fn find_evictable_connection(&self) -> Option<ConnectionTuple>

fn enqueue_rst_config(&mut self, tuple: ConnectionTuple, cfg: RstConfig)

fn enqueue_rst<T: NetworkBytes + Debug>( &mut self, tuple: ConnectionTuple, s: &TcpSegment<'_, T> )

pub fn write_next_packet( &mut self, buf: &mut [u8] ) -> Result<(Option<NonZeroUsize>, WriteEvent), WriteNextError>

Attempts to write one packet, from either the RST queue or one of the existing endpoints, to buf.

On success, the function returns a pair containing an Option<NonZeroUsize> and a WriteEvent. The options represents how many bytes have been written to buf, or that no packet can be send presently (when equal to None). The WriteEvent describes whether any noteworthy state changes are associated with the write.

pub fn next_segment_status(&self) -> NextSegmentStatus

Describes the status of the next segment to be sent by the handler.