Struct vmm::Vmm

pub struct Vmm {
    pub(crate) events_observer: Option<Stdin>,
    pub(crate) instance_info: InstanceInfo,
    pub(crate) shutdown_exit_code: Option<FcExitCode>,
    pub(crate) vm: Vm,
    pub(crate) guest_memory: GuestMemoryMmap,
    pub(crate) uffd: Option<Uffd>,
    pub(crate) vcpus_handles: Vec<VcpuHandle>,
    pub(crate) vcpus_exit_evt: EventFd,
    pub(crate) mmio_device_manager: MMIODeviceManager,
    pub(crate) pio_device_manager: PortIODeviceManager,
}

Contains the state and associated methods required for the Firecracker VMM.

Fields
events_observer: Option<Stdin>instance_info: InstanceInfoshutdown_exit_code: Option<FcExitCode>vm: Vmguest_memory: GuestMemoryMmapuffd: Option<Uffd>vcpus_handles: Vec<VcpuHandle>vcpus_exit_evt: EventFdmmio_device_manager: MMIODeviceManagerpio_device_manager: PortIODeviceManager
Implementations

impl Vmm

pub fn version(&self) -> String

Gets Vmm version.

pub fn instance_info(&self) -> InstanceInfo

Gets Vmm instance info.

pub fn shutdown_exit_code(&self) -> Option<FcExitCode>

Provides the Vmm shutdown exit code if there is one.

pub fn get_bus_device( &self, device_type: DeviceType, device_id: &str ) -> Option<&Mutex<BusDevice>>

Gets the specified bus device.

pub fn start_vcpus( &mut self, vcpus: Vec<Vcpu>, vcpu_seccomp_filter: Arc<BpfProgram> ) -> Result<(), StartVcpusError>

Starts the microVM vcpus.

Errors

When:

[vmm::VmmEventsObserver::on_vmm_boot] errors.

[vmm::vstate::vcpu::Vcpu::start_threaded] errors.

pub fn resume_vm(&mut self) -> Result<(), VmmError>

Sends a resume command to the vCPUs.

pub fn pause_vm(&mut self) -> Result<(), VmmError>

Sends a pause command to the vCPUs.

pub fn guest_memory(&self) -> &GuestMemoryMmap

Returns a reference to the inner GuestMemoryMmap object.

pub fn emulate_serial_init(&self) -> Result<(), EmulateSerialInitError>

Sets RDA bit in serial console

pub fn send_ctrl_alt_del(&mut self) -> Result<(), VmmError>

Injects CTRL+ALT+DEL keystroke combo in the i8042 device.

pub fn save_state( &mut self, vm_info: &VmInfo ) -> Result<MicrovmState, MicrovmStateError>

Saves the state of a paused Microvm.

pub(crate) fn save_vcpu_states( &mut self ) -> Result<Vec<VcpuState>, MicrovmStateError>

pub fn dump_cpu_config( &mut self ) -> Result<Vec<CpuConfiguration>, DumpCpuConfigError>

Dumps CPU configuration.

pub fn get_dirty_bitmap(&self) -> Result<DirtyBitmap, VmmError>

Retrieves the KVM dirty bitmap for each of the guest’s memory regions.

pub fn set_dirty_page_tracking(&mut self, enable: bool) -> Result<(), VmmError>

Enables or disables KVM dirty page tracking.

pub fn update_block_device_path( &mut self, drive_id: &str, path_on_host: String ) -> Result<(), VmmError>

Updates the path of the host file backing the emulated block device with id drive_id. We update the disk image on the device and its virtio configuration.

pub fn update_block_rate_limiter( &mut self, drive_id: &str, rl_bytes: BucketUpdate, rl_ops: BucketUpdate ) -> Result<(), VmmError>

Updates the rate limiter parameters for block device with drive_id id.

pub fn update_vhost_user_block_config( &mut self, drive_id: &str ) -> Result<(), VmmError>

Updates the rate limiter parameters for block device with drive_id id.

pub fn update_net_rate_limiters( &mut self, net_id: &str, rx_bytes: BucketUpdate, rx_ops: BucketUpdate, tx_bytes: BucketUpdate, tx_ops: BucketUpdate ) -> Result<(), VmmError>

Updates the rate limiter parameters for net device with net_id id.

pub fn balloon_config(&self) -> Result<BalloonConfig, BalloonError>

Returns a reference to the balloon device if present.

pub fn latest_balloon_stats(&self) -> Result<BalloonStats, BalloonError>

Returns the latest balloon statistics if they are enabled.

pub fn update_balloon_config( &mut self, amount_mib: u32 ) -> Result<(), BalloonError>

Updates configuration for the balloon device target size.

pub fn update_balloon_stats_config( &mut self, stats_polling_interval_s: u16 ) -> Result<(), BalloonError>

Updates configuration for the balloon device as described in balloon_stats_update.

pub fn stop(&mut self, exit_code: FcExitCode)

Signals Vmm to stop and exit.