Struct vmm::vstate::vm::Vm

pub struct Vm {
    fd: VmFd,
    max_memslots: usize,
    pub kvm_cap_modifiers: Vec<KvmCapability>,
    supported_cpuid: CpuId,
    msrs_to_save: MsrList,
}

A wrapper around creating and using a VM.

Fields
fd: VmFdmax_memslots: usizekvm_cap_modifiers: Vec<KvmCapability>

Additional capabilities that were specified in cpu template.

supported_cpuid: CpuIdmsrs_to_save: MsrList
Implementations

impl Vm

Contains Vm functions that are usable across CPU architectures

pub fn new(kvm_cap_modifiers: Vec<KvmCapability>) -> Result<Self, VmError>

Constructs a new Vm using the given Kvm instance.

fn combine_capabilities(kvm_cap_modifiers: &[KvmCapability]) -> Vec<u32>

fn check_capabilities(kvm: &Kvm, capabilities: &[u32]) -> Result<(), u32>

pub fn memory_init( &self, guest_mem: &GuestMemoryMmap, track_dirty_pages: bool ) -> Result<(), VmError>

Initializes the guest memory.

pub(crate) fn set_kvm_memory_regions( &self, guest_mem: &GuestMemoryMmap, track_dirty_pages: bool ) -> Result<(), VmError>

pub fn fd(&self) -> &VmFd

Gets a reference to the kvm file descriptor owned by this VM.

impl Vm

const DEFAULT_CAPABILITIES: [u32; 14] = _

pub fn supported_cpuid(&self) -> &CpuId

Returns a ref to the supported CpuId for this Vm.

pub fn msrs_to_save(&self) -> &MsrList

Returns a ref to the list of serializable MSR indices.

pub fn restore_state( &mut self, state: &VmState ) -> Result<(), RestoreStateError>

Restores the KVM VM state.

Errors

When:

[kvm_ioctls::VmFd::set_pit] errors.

[kvm_ioctls::VmFd::set_clock] errors.

[kvm_ioctls::VmFd::set_irqchip] errors.

[kvm_ioctls::VmFd::set_irqchip] errors.

[kvm_ioctls::VmFd::set_irqchip] errors.

pub fn setup_irqchip(&self) -> Result<(), VmError>

Creates the irq chip and an in-kernel device model for the PIT.

pub fn save_state(&self) -> Result<VmState, VmError>

Saves and returns the Kvm Vm state.